changed svg on newcase page
