+++ src/pages/NewCase.js
@@ -7,7 +7,7 @@ import CanvasArea from "../components/CanvasArea";
 import NodeList from "../components/NodeList";
 import SideNav from "../components/Sidebar";
 import Tool from "../components/Tool";
-import {AiOutlineCloseCircle} from "react-icons/ai";
+// import {AiOutlineCloseCircle} from "react-icons/ai";
 
 
 
@@ -31,7 +31,7 @@ const NewCase = () => {
   const {ROUTES} = constants;
   const [isOpen, setIsOpen] = useState(false);
   const [isshow, setIsshow] = useState(false);
-  const [isshowDiv, setIsshowDiv] = useState(true);
+  // const [isshowDiv, setIsshowDiv] = useState(true);
   const [isHovering, setIsHovering] = useState(false);
   const navigate = useNavigate();
   const searchRef = useRef();
@@ -71,9 +71,9 @@ const NewCase = () => {
     setIsshow(!isshow);
   };
 
-  const handleClose = () => {
-    setIsshowDiv(false);
-  };
+  // const handleClose = () => {
+  //   setIsshowDiv(false);
+  // };
 
   const handleMouseOver = () => {
     setIsHovering(true);
@@ -104,198 +104,15 @@ const NewCase = () => {
           <div>
 
           <button className='btn-icon'onClick={opendrawer}>
-          <svg width="23" height="23" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path  d="M21.9255 12.2275C21.9255 18.0265 17.2245 22.7275 11.4255 22.7275C5.62653 22.7275 0.925537 18.0265 0.925537 12.2275C0.925537 6.42855 5.62653 1.72754 11.4255 1.72754C17.2245 1.72754 21.9255 6.42855 21.9255 12.2275Z" fill="black" fill-opacity="0.1"/>
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M11.4255 0.852539C11.4255 0.852539 13.7392 0.852539 15.8535 1.7468C15.8535 1.7468 17.895 2.61028 19.4689 4.1842C19.4689 4.1842 21.0428 5.75812 21.9063 7.79961C21.9063 7.79961 22.8005 9.91389 22.8005 12.2275C22.8005 12.2275 22.8005 14.5412 21.9063 16.6555C21.9063 16.6555 21.0428 18.697 19.4689 20.2709C19.4689 20.2709 17.895 21.8448 15.8535 22.7083C15.8535 22.7083 13.7392 23.6025 11.4255 23.6025C11.4255 23.6025 9.11189 23.6025 6.99761 22.7083C6.99761 22.7083 4.95612 21.8448 3.3822 20.2709C3.3822 20.2709 1.80827 18.697 0.944801 16.6555C0.944801 16.6555 0.0505371 14.5412 0.0505371 12.2275C0.0505371 12.2275 0.0505371 9.91388 0.944801 7.79961C0.944801 7.79961 1.80828 5.75812 3.3822 4.1842C3.3822 4.1842 4.95612 2.61028 6.99761 1.7468C6.99761 1.7468 9.11188 0.852539 11.4255 0.852539ZM11.4255 2.60254C11.4255 2.60254 9.46676 2.60254 7.67932 3.35856C7.67932 3.35856 5.9522 4.08907 4.61964 5.42164C4.61964 5.42164 3.28707 6.7542 2.55656 8.48132C2.55656 8.48132 1.80054 10.2688 1.80054 12.2275C1.80054 12.2275 1.80054 14.1863 2.55656 15.9738C2.55656 15.9738 3.28707 17.7009 4.61963 19.0334C4.61963 19.0334 5.9522 20.366 7.67932 21.0965C7.67932 21.0965 9.46676 21.8525 11.4255 21.8525C11.4255 21.8525 13.3843 21.8525 15.1718 21.0965C15.1718 21.0965 16.8989 20.366 18.2314 19.0334C18.2314 19.0334 19.564 17.7009 20.2945 15.9738C20.2945 15.9738 21.0505 14.1863 21.0505 12.2275C21.0505 12.2275 21.0505 10.2688 20.2945 8.48132C20.2945 8.48132 19.564 6.7542 18.2314 5.42164C18.2314 5.42164 16.8989 4.08907 15.1718 3.35856C15.1718 3.35856 13.3843 2.60254 11.4255 2.60254Z" fill="#1C1C1C"/>
-            <path d="M12.738 17.915C12.738 18.6399 12.1504 19.2275 11.4255 19.2275C10.7007 19.2275 10.113 18.6399 10.113 17.915C10.113 17.1902 10.7007 16.6025 11.4255 16.6025C12.1504 16.6025 12.738 17.1902 12.738 17.915Z" fill="#1C1C1C"/>
-            <path d="M8.15163 7.85258C7.48804 8.84571 7.48804 10.0401 7.48804 10.0401C7.48804 10.5234 7.87979 10.9151 8.36304 10.9151C8.84629 10.9151 9.23804 10.5234 9.23804 10.0401C9.23804 9.37656 9.6067 8.82482 9.6067 8.82482C9.97536 8.27308 10.5884 8.01915 10.5884 8.01915C11.2015 7.76521 11.8523 7.89467 11.8523 7.89467C12.5031 8.02412 12.9723 8.49334 12.9723 8.49334C13.4415 8.96255 13.571 9.61337 13.571 9.61337C13.7005 10.2642 13.4465 10.8773 13.4465 10.8773C13.1926 11.4903 12.6408 11.859 12.6408 11.859C12.0891 12.2276 11.4255 12.2276 11.4255 12.2276C10.9423 12.2276 10.5505 12.6194 10.5505 13.1026V13.9776C10.5505 14.4609 10.9423 14.8526 11.4255 14.8526C11.9088 14.8526 12.3005 14.4609 12.3005 13.9776V13.8826C12.9957 13.7266 13.6131 13.314 13.6131 13.314C14.6062 12.6505 15.0633 11.5469 15.0633 11.5469C15.5204 10.4434 15.2874 9.27196 15.2874 9.27196C15.0544 8.10049 14.2098 7.2559 14.2098 7.2559C13.3652 6.41131 12.1937 6.17829 12.1937 6.17829C11.0222 5.94527 9.91872 6.40236 9.91872 6.40236C8.81521 6.85945 8.15163 7.85258 8.15163 7.85258Z" fill="#1C1C1C"/>
-    </svg>    
+          <img src={bookmark} alt="bookmark" />
 </button>
 {/* HELP BUTTON DATA--------------------------------------------------------------------------------- */}
-  {isshowDiv && (
-  <div>
-{isOpen && (
-                <span className="help">
-                  <div onClick={handleClose} className="close-btn">
-                  <AiOutlineCloseCircle/>
-                  </div>
-                  <h2>Search Guidelines</h2>
-                  <hr />
-                  <h4>Keyword Search</h4>
-                  <p>Friday Intel offers full-text search and retrieves recent darkweb <br /> content that includes the search keyword.</p>
-                  <p>Just as search engines, put Keybord to search</p>
-                  <br />
-                  
-<p>If there are too many results to study, try a more specific keyword search to get more relevant results. To find more specific results, use the search operator.</p>
-<p>Just as search engines, put Keybord to search</p>
-<h1>Search Operator</h1>
-<p>Search operators are special commands that enable advanced full text</p>
-<div className="parent">
-<div className="Diff1">
-<h4>Operator &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Description</h4>
-<hr />
-<p>"" &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;The search results are exact matches &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;to the keywords in""</p>
-<hr />
-<p>AND &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Search results include both X and Y</p>
-<hr />
-<p>OR &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Search results include X and Y</p>
-<hr />
-<p>NOT &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Search results include X but exclude &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; results contain Y</p>
-<hr />
-<p>Sites:&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Filtered site search results &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Support for the darkweb/surfaceweb website &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;A site ; must be a full domain address &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;(perfect match), for example Friday Intel &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; NOT  inurl:xxxxxxxxxxxxx.onion</p>
-<hr />
-<p>inurl: &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Search result in filtered URL  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Support for darkweb/surfaceweb  URL  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;should be match with specific pattern &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; (Like match) e.g.Friday Intel NOT &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;inurl http://xxxxxxxxx.com</p>
-</div>
-</div>
-<h4>Indicator  Search</h4>
-<p>The purpose of indicator search is to extract information that is linked to the searched keyword. The keyword or source of the keyword could be used to generate a list of analysis results. The Monography of Friday Intel button displays a list of indicators that you can use. To search, click an indicator and enter an exact keyword for a perfect match.
-</p>
-<p>Indicator Remarks:</p>
-<h4>Indicator &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Remarks  </h4>
-<hr />
-<p>tor/i2p:&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Retrieves all information obtained 
-from &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; a dark web site and can perform 
-total &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;domain analysis.</p>
-<hr />
-<p>torurl/i2purl: &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Retrieves all information obtained 
-from a   &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;dark web URL and performs a 
-total analysis   &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;on the searched URL</p>
-<hr />
-<p>domain: &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; It is useful for tracking any mentions 
-of  &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; a searched site on the dark web.</p>
-<hr />
-<p>hash: &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Uses md5/sha1/sha2 to retrieve any 
-files.</p>
-<div className="shape3"></div>
-</span>
-)}
-</div>
-)}
+   {isOpen && <DrawerInfo />}
 </div>
           <div className='searchbar-box'>
-          <svg className='searchbar-logo' width="27" height="24" viewBox="0 0 27 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M18.2852 9.3595C18.1709 9.82532 17.1123 9.88437 16.568 9.75315C16.0238 9.62194 15.2496 9.16137 15.3625 8.69555C15.4755 8.22973 16.6225 7.16425 17.1681 7.28628C17.7137 7.40831 18.3995 8.89237 18.2852 9.3595Z" fill="#00FFFF" />
-            <path d="M15.8387 10.2177L15.8673 10.173C15.8782 10.1599 15.8891 10.1455 15.9013 10.1324C15.9284 10.1065 15.9591 10.0844 15.9925 10.0668C16.0605 10.0313 16.1381 10.0143 16.2061 9.98541C16.2394 9.97202 16.2709 9.95484 16.3 9.93423L16.3381 9.90012C16.3493 9.88756 16.3597 9.87442 16.3694 9.86075C16.4093 9.80849 16.4321 9.74586 16.4347 9.68098C16.431 9.61382 16.4172 9.54753 16.3939 9.48416C16.3834 9.4491 16.377 9.413 16.3748 9.37656C16.3741 9.33922 16.3797 9.30201 16.3912 9.26634C16.4149 9.19501 16.4587 9.13142 16.5177 9.08264C16.5812 9.03481 16.6545 9.00039 16.7327 8.9816C16.8021 8.96639 16.8674 8.93734 16.9245 8.89631C16.975 8.85409 17.013 8.79977 17.0347 8.73885C17.0545 8.67731 17.0545 8.61144 17.0347 8.5499C17.011 8.48182 16.9991 8.41045 16.9994 8.33864C17.0056 8.26584 17.034 8.19645 17.081 8.13919C17.093 8.12502 17.1057 8.11144 17.1191 8.09851C17.1327 8.08539 17.149 8.07358 17.164 8.06177C17.1971 8.04058 17.2327 8.0234 17.2701 8.0106C17.3436 7.98567 17.4212 7.97386 17.4892 7.94368C17.5551 7.91534 17.6117 7.86992 17.6525 7.81246L17.6253 7.85576C17.613 7.8702 17.6021 7.88463 17.5899 7.89775C17.5631 7.92306 17.5329 7.94469 17.5001 7.96205C17.432 7.99879 17.3531 8.01585 17.2865 8.04471C17.2528 8.05729 17.2211 8.07453 17.1926 8.09589L17.1545 8.12869C17.1429 8.14091 17.1324 8.15408 17.1232 8.16806C17.083 8.22032 17.0598 8.28287 17.0565 8.34782C17.0607 8.41504 17.0749 8.48133 17.0987 8.54465C17.1088 8.57976 17.1147 8.61585 17.1164 8.65225C17.1182 8.68955 17.1131 8.72687 17.1014 8.76247C17.0771 8.83357 17.0334 8.89703 16.9749 8.94617C16.9108 8.99403 16.8372 9.02844 16.7585 9.04721C16.6888 9.06228 16.6233 9.09183 16.5667 9.13381C16.5167 9.17591 16.4787 9.22967 16.4565 9.28996C16.4461 9.31953 16.441 9.35059 16.4415 9.38181C16.4421 9.41458 16.4467 9.44716 16.4551 9.47891C16.4787 9.54706 16.4911 9.61834 16.4918 9.69017C16.4852 9.76309 16.4563 9.83248 16.4088 9.88962C16.3969 9.90379 16.3842 9.91737 16.3708 9.93029C16.3567 9.94357 16.3417 9.95585 16.3259 9.96704C16.2934 9.98847 16.2582 10.0057 16.2211 10.0182C16.1462 10.0445 16.0687 10.055 16.0007 10.0864C15.9349 10.1145 15.8787 10.16 15.8387 10.2177Z" fill="#2FBACE" />
-            <path d="M15.5122 9.88174L15.5517 9.84894L15.5979 9.82007C15.6308 9.80299 15.6661 9.79061 15.7027 9.78333C15.7771 9.77214 15.852 9.76469 15.9272 9.76103C15.9624 9.75672 15.9971 9.7488 16.0306 9.73741L16.0769 9.7151C16.0917 9.70714 16.1058 9.69792 16.119 9.68754C16.1717 9.65049 16.2127 9.60007 16.2374 9.54189C16.2549 9.48082 16.2632 9.41761 16.2619 9.35425C16.2638 9.31961 16.2697 9.28528 16.2796 9.25191C16.2904 9.2176 16.3069 9.18524 16.3286 9.15612C16.3753 9.09884 16.4361 9.05373 16.5055 9.0249C16.5796 8.99714 16.6586 8.98377 16.7381 8.98554C16.8085 8.98832 16.8788 8.97762 16.9449 8.95404C17.0046 8.92791 17.0569 8.8883 17.0973 8.83857C17.1347 8.78928 17.1561 8.73035 17.1586 8.6693C17.158 8.60144 17.1695 8.53398 17.1926 8.46986C17.2209 8.40491 17.2682 8.34929 17.3286 8.30977C17.3441 8.29892 17.3605 8.28927 17.3776 8.2809C17.3939 8.27303 17.413 8.26647 17.4307 8.25991C17.4675 8.24768 17.506 8.24061 17.545 8.23891C17.6225 8.23891 17.6987 8.23891 17.7708 8.22973C17.8426 8.21946 17.9102 8.19092 17.9668 8.14706L17.926 8.18118C17.9115 8.19112 17.8965 8.20032 17.8811 8.20873C17.848 8.22554 17.8128 8.23791 17.7763 8.24548C17.7014 8.26253 17.6225 8.2586 17.5504 8.26909C17.5151 8.27292 17.4804 8.28085 17.447 8.29271L17.4007 8.31371C17.3866 8.32214 17.373 8.33134 17.3599 8.34126C17.3061 8.37775 17.2649 8.42894 17.2416 8.48823C17.2236 8.54879 17.2149 8.61156 17.2157 8.67455C17.2143 8.70915 17.2088 8.74349 17.1994 8.7769C17.1887 8.81149 17.1716 8.84394 17.149 8.87269C17.1028 8.92987 17.0425 8.97498 16.9735 9.00391C16.8988 9.03126 16.8193 9.04419 16.7395 9.04196C16.6695 9.03907 16.5996 9.04978 16.534 9.07345C16.474 9.09903 16.4216 9.13873 16.3816 9.18892C16.3619 9.21288 16.3468 9.23997 16.3367 9.26896C16.3283 9.29857 16.3229 9.32889 16.3204 9.3595C16.3197 9.42695 16.3077 9.49385 16.285 9.55764C16.2564 9.62277 16.2092 9.67873 16.149 9.71904C16.134 9.72953 16.1163 9.73741 16.1013 9.74659L16.0469 9.7689C16.0097 9.77992 15.9714 9.78696 15.9326 9.78989C15.8551 9.78989 15.7802 9.78989 15.7068 9.79777C15.6353 9.80864 15.5682 9.83763 15.5122 9.88174Z" fill="#2FBACE" />
-            <path d="M14.8536 9.86994L14.9162 9.83451C14.9394 9.82663 14.9611 9.81614 14.9856 9.80826C15.0341 9.79325 15.0844 9.78486 15.1353 9.78333C15.2401 9.78333 15.3421 9.8017 15.4414 9.80433C15.4897 9.80767 15.5381 9.80502 15.5857 9.79645L15.6537 9.7794C15.6756 9.77201 15.6969 9.76324 15.7176 9.75315C15.8 9.71572 15.8702 9.65717 15.9204 9.58388C15.9619 9.50893 15.9917 9.42845 16.0088 9.34507C16.022 9.29977 16.0398 9.25582 16.0619 9.21385C16.0873 9.17168 16.1195 9.13361 16.1571 9.10101C16.2351 9.03473 16.3284 8.98748 16.4293 8.96323C16.5353 8.9435 16.6443 8.9435 16.7504 8.96323C16.8443 8.98089 16.9408 8.98089 17.0347 8.96323C17.122 8.94337 17.2032 8.90379 17.2715 8.84776C17.3373 8.79275 17.3846 8.72013 17.4076 8.63913C17.4279 8.55059 17.4633 8.46591 17.5123 8.3885C17.5724 8.30935 17.6538 8.24762 17.7477 8.21005C17.7711 8.20017 17.7952 8.19184 17.8198 8.18512L17.8974 8.16806C17.9503 8.16281 18.0036 8.16281 18.0566 8.16806C18.16 8.17855 18.2593 8.20611 18.3586 8.20742C18.4577 8.21085 18.5559 8.18865 18.643 8.14313L18.5791 8.17855L18.511 8.2048C18.4622 8.2203 18.4113 8.2287 18.36 8.22973C18.2566 8.22973 18.1532 8.21267 18.0538 8.20873C18.0056 8.20609 17.9572 8.20873 17.9096 8.21661L17.8416 8.23367C17.8212 8.24285 17.7981 8.24941 17.7776 8.25991C17.697 8.29698 17.628 8.35403 17.5776 8.42524C17.5354 8.49957 17.5051 8.57965 17.4878 8.66274C17.4759 8.70844 17.458 8.75251 17.4348 8.79396C17.4098 8.83609 17.3781 8.87417 17.3409 8.90681C17.2621 8.97059 17.1689 9.01555 17.0688 9.03802C16.9608 9.06692 16.8477 9.07407 16.7368 9.05902C16.6433 9.0414 16.5472 9.0414 16.4537 9.05902C16.3664 9.07936 16.2852 9.11938 16.217 9.1758C16.1841 9.20209 16.156 9.23359 16.134 9.26896C16.1132 9.30523 16.0963 9.34349 16.0837 9.38312C16.0633 9.47165 16.0279 9.55633 15.9789 9.63375C15.9196 9.71309 15.8386 9.7749 15.7449 9.8122C15.7209 9.82188 15.6964 9.8302 15.6714 9.83713C15.6459 9.8442 15.62 9.8499 15.5938 9.85419C15.5414 9.86076 15.4884 9.86076 15.436 9.85419C15.3312 9.84369 15.2333 9.81614 15.1339 9.81482C15.0372 9.80975 14.9406 9.82872 14.8536 9.86994Z" fill="#2FBACE" />
-            <path d="M15.104 9.32145L15.1611 9.30702C15.1802 9.30702 15.1992 9.30702 15.2197 9.29914C15.2598 9.29392 15.3006 9.29392 15.3408 9.29914C15.421 9.31358 15.4945 9.349 15.5693 9.36738C15.6055 9.37835 15.643 9.38496 15.6809 9.38706H15.7381C15.7557 9.38706 15.7748 9.38706 15.7925 9.37787C15.8628 9.3649 15.9284 9.33461 15.983 9.28996C16.0318 9.24176 16.073 9.18695 16.1054 9.12725C16.1262 9.09634 16.1499 9.06735 16.1762 9.04065C16.2053 9.01355 16.2383 8.99058 16.2741 8.97241C16.3495 8.93665 16.4323 8.91781 16.5163 8.9173C16.6009 8.92125 16.684 8.94038 16.7613 8.97373C16.8292 9.0045 16.9026 9.02233 16.9776 9.02621C17.0493 9.02628 17.12 9.01058 17.1844 8.98029C17.2458 8.95113 17.2973 8.90559 17.3327 8.84907C17.3688 8.78622 17.4147 8.72912 17.4688 8.6798C17.532 8.63089 17.6076 8.59918 17.6878 8.58795C17.7085 8.58404 17.7294 8.58141 17.7504 8.58008H17.813C17.8542 8.58463 17.8948 8.59343 17.9341 8.60632C18.0117 8.63256 18.0811 8.6693 18.1573 8.68767C18.2331 8.70668 18.3127 8.70668 18.3886 8.68767L18.3328 8.70211L18.2729 8.71129C18.2328 8.71391 18.1925 8.71126 18.1532 8.70342C18.0715 8.68899 17.9981 8.65487 17.9246 8.63519C17.8879 8.62462 17.8499 8.61845 17.8117 8.61682C17.7931 8.61551 17.7744 8.61551 17.7559 8.61682C17.7375 8.61806 17.7193 8.62069 17.7014 8.62469C17.6306 8.63716 17.5644 8.66749 17.5096 8.71261C17.4608 8.7608 17.4195 8.81561 17.3871 8.87531C17.3668 8.9065 17.3431 8.93552 17.3164 8.96192C17.2876 8.98941 17.2545 9.01243 17.2184 9.03015C17.1432 9.06631 17.0603 9.08517 16.9762 9.08526C16.8916 9.08175 16.8085 9.0626 16.7313 9.02884C16.6636 8.9987 16.5907 8.9809 16.5163 8.97635C16.4443 8.97675 16.3732 8.99243 16.3082 9.02228C16.2777 9.0371 16.2494 9.05562 16.2238 9.07739C16.2 9.10116 16.1785 9.12708 16.1599 9.15481C16.1235 9.21706 16.0776 9.27369 16.0238 9.32276C15.9597 9.3713 15.8838 9.40336 15.8034 9.41593C15.783 9.41593 15.7625 9.41593 15.7421 9.42249H15.6796C15.6379 9.41779 15.5969 9.40899 15.5571 9.39624C15.4795 9.37131 15.4115 9.33326 15.3353 9.31489C15.2588 9.29886 15.1794 9.30112 15.104 9.32145Z" fill="#2FBACE" />
-            <path d="M14.4754 8.91599H14.5543C14.5815 8.91599 14.6087 8.91599 14.6346 8.92386C14.6883 8.93194 14.7405 8.94741 14.7897 8.96979C14.889 9.01178 14.9707 9.0787 15.0618 9.12594C15.1049 9.14985 15.1505 9.16921 15.1979 9.18367L15.2727 9.19811C15.2976 9.19945 15.3226 9.19945 15.3476 9.19811C15.4455 9.2029 15.5431 9.18443 15.6319 9.14431C15.7144 9.09802 15.79 9.04107 15.8564 8.97504C15.8957 8.94105 15.9386 8.91114 15.9843 8.88581C16.033 8.85962 16.0852 8.84019 16.1394 8.82808C16.2505 8.8038 16.3658 8.8038 16.4769 8.82808C16.5847 8.85837 16.6858 8.90776 16.7749 8.97373C16.8518 9.03337 16.9401 9.07794 17.0347 9.10494C17.1302 9.12527 17.2291 9.12527 17.3246 9.10494C17.4168 9.08525 17.5021 9.04275 17.5722 8.9816C17.6473 8.91502 17.7327 8.86013 17.8253 8.81889C17.9273 8.77661 18.0384 8.75904 18.1491 8.76772C18.177 8.76928 18.2047 8.77279 18.2321 8.77821C18.2593 8.78346 18.2865 8.79002 18.3137 8.7979C18.3657 8.81575 18.4154 8.83907 18.462 8.86744C18.5532 8.92386 18.628 8.9921 18.7192 9.03802C18.8128 9.08525 18.917 9.10958 19.0226 9.10888H18.9423L18.8621 9.10101C18.8088 9.09293 18.757 9.07746 18.7083 9.05508C18.6076 9.01309 18.526 8.94748 18.4362 8.90024C18.3932 8.87611 18.3476 8.85674 18.3001 8.84251C18.252 8.82856 18.202 8.82149 18.1518 8.82152C18.0535 8.81749 17.9556 8.83592 17.8661 8.87531C17.7841 8.92226 17.7086 8.97915 17.6416 9.04458C17.6026 9.07869 17.5602 9.10902 17.515 9.13512C17.4665 9.16122 17.4142 9.18024 17.3599 9.19155C17.2444 9.21718 17.1242 9.21673 17.0089 9.19023C16.901 9.16103 16.8002 9.11153 16.7123 9.04458C16.6347 8.98518 16.546 8.94064 16.451 8.91337C16.3556 8.89304 16.2567 8.89304 16.1612 8.91337C16.1139 8.92194 16.0681 8.93696 16.0251 8.95798C15.9843 8.97971 15.9459 9.00567 15.9109 9.0354C15.8377 9.10465 15.7542 9.16303 15.6632 9.2086C15.5615 9.25168 15.4501 9.26929 15.3394 9.25978C15.3108 9.25978 15.2836 9.25322 15.255 9.25059C15.2279 9.24499 15.2011 9.23798 15.1748 9.2296C15.1226 9.2123 15.0728 9.18895 15.0264 9.16005C14.9353 9.10363 14.8604 9.02884 14.7679 8.98947C14.678 8.94264 14.5776 8.91743 14.4754 8.91599Z" fill="#2FBACE" />
-            <path d="M15.0945 8.53415L15.153 8.54727C15.172 8.54727 15.1924 8.55908 15.2115 8.56696C15.2479 8.58296 15.2822 8.60324 15.3135 8.62731C15.371 8.67813 15.4256 8.73202 15.4768 8.78871C15.5038 8.81353 15.5335 8.83552 15.5653 8.85432C15.5819 8.86326 15.5992 8.87115 15.617 8.87794C15.6341 8.88542 15.6519 8.89157 15.67 8.89631C15.7425 8.91645 15.8188 8.92049 15.8932 8.90812C15.9653 8.8892 16.0347 8.86188 16.1 8.82676C16.1368 8.80963 16.175 8.79558 16.2143 8.78477C16.2561 8.775 16.2992 8.77102 16.3422 8.77296C16.429 8.77493 16.5141 8.79602 16.5912 8.83464C16.6629 8.87501 16.7257 8.9285 16.7762 8.9921C16.8169 9.05104 16.8716 9.09977 16.9357 9.13423C16.9998 9.16869 17.0716 9.18789 17.1449 9.19023C17.2184 9.19256 17.2914 9.17723 17.3572 9.14562C17.428 9.10935 17.5035 9.08243 17.5817 9.06558C17.6658 9.05407 17.7516 9.06171 17.8321 9.08788C17.8721 9.10062 17.9104 9.11779 17.9464 9.13906C17.9802 9.16088 18.0112 9.18642 18.0389 9.21516C18.0933 9.27159 18.1314 9.33457 18.1886 9.38443C18.2475 9.43531 18.3185 9.47137 18.3954 9.48941L18.3355 9.47629C18.3165 9.46972 18.2974 9.46448 18.2784 9.4566C18.241 9.44141 18.2062 9.42108 18.1749 9.39624C18.1124 9.34769 18.0675 9.28602 18.0117 9.23616C17.9853 9.21031 17.9555 9.18783 17.9232 9.16924C17.9069 9.16137 17.8906 9.15218 17.8729 9.14562C17.8552 9.13906 17.8375 9.1325 17.8185 9.12856C17.7462 9.10727 17.6697 9.10278 17.5953 9.11544C17.5233 9.13464 17.4539 9.16195 17.3885 9.19679C17.352 9.21449 17.3137 9.22856 17.2742 9.23878C17.2323 9.24847 17.1894 9.25288 17.1463 9.25191C17.0595 9.2487 16.9745 9.2272 16.8973 9.18892C16.825 9.14933 16.762 9.09572 16.7123 9.03146C16.6694 8.97406 16.6148 8.92581 16.5517 8.88975C16.4858 8.85722 16.4134 8.83886 16.3395 8.83595C16.3034 8.83398 16.2673 8.83663 16.232 8.84382C16.1964 8.85249 16.1618 8.86435 16.1286 8.87925C16.0577 8.91532 15.9822 8.94223 15.9041 8.95929C15.8195 8.97101 15.7333 8.96337 15.6523 8.93699C15.6319 8.93699 15.6129 8.92255 15.5938 8.91599L15.538 8.88581C15.5039 8.86439 15.4728 8.83881 15.4455 8.80971C15.3989 8.75112 15.3489 8.69505 15.2959 8.64175C15.2392 8.59014 15.1699 8.55313 15.0945 8.53415Z" fill="#2FBACE" />
-            <path d="M14.9013 8.12344L14.9679 8.14969L15.0319 8.1838C15.0715 8.20948 15.1073 8.2404 15.138 8.27565C15.2006 8.3452 15.2373 8.42655 15.2904 8.4961C15.3154 8.53162 15.3452 8.56383 15.3788 8.59189C15.3965 8.60501 15.4129 8.61813 15.4319 8.62994C15.4513 8.64144 15.4713 8.65195 15.4918 8.66143C15.5737 8.70086 15.6642 8.72109 15.7557 8.72048C15.8485 8.71383 15.9399 8.69575 16.0279 8.66668C16.0764 8.65412 16.1261 8.64577 16.1762 8.64175C16.2286 8.63805 16.2813 8.64159 16.3327 8.65225C16.4323 8.66969 16.5261 8.7102 16.606 8.77031C16.6859 8.83042 16.7494 8.90833 16.7912 8.99735C16.8267 9.07094 16.8777 9.13662 16.9409 9.19023C17.0119 9.24173 17.0936 9.27796 17.1803 9.29652C17.2676 9.31483 17.3583 9.31213 17.4443 9.28865C17.5383 9.26003 17.6357 9.24327 17.7341 9.23878C17.8369 9.24182 17.9375 9.26788 18.028 9.31489C18.05 9.3265 18.0713 9.3392 18.0919 9.35294C18.1124 9.36738 18.1314 9.38312 18.1505 9.39887C18.1853 9.43066 18.2155 9.46685 18.2403 9.50647C18.2906 9.58388 18.3192 9.66393 18.3763 9.73347C18.4335 9.80574 18.5089 9.86267 18.5954 9.8988L18.5274 9.87125L18.4648 9.83713C18.4247 9.81151 18.3885 9.78059 18.3573 9.74528C18.2947 9.67574 18.2579 9.59438 18.2062 9.52484C18.1804 9.48992 18.1508 9.4578 18.1178 9.42905C18.1001 9.41593 18.0824 9.4028 18.0634 9.391C18.0443 9.37919 18.0253 9.36869 18.0035 9.3595C17.9218 9.32082 17.832 9.30062 17.7409 9.30046C17.6483 9.30757 17.5569 9.32564 17.4688 9.35425C17.4203 9.36733 17.3706 9.37569 17.3205 9.37919C17.2681 9.38359 17.2153 9.38005 17.164 9.36869C17.0621 9.34918 16.9663 9.30743 16.8837 9.24666C16.8082 9.18588 16.747 9.11028 16.7041 9.0249C16.67 8.94662 16.6162 8.87779 16.5475 8.82463C16.4788 8.77147 16.3975 8.73567 16.3109 8.72048C16.2664 8.71022 16.2205 8.70624 16.1748 8.70867C16.129 8.71189 16.0835 8.71847 16.0388 8.72835C15.9453 8.75709 15.8483 8.77385 15.7503 8.77821C15.6475 8.77537 15.5468 8.7493 15.4564 8.70211C15.4339 8.69069 15.4121 8.67798 15.3911 8.66406C15.3707 8.65032 15.3512 8.63543 15.3326 8.61944C15.2981 8.58685 15.2679 8.55027 15.2428 8.51053C15.1938 8.43443 15.1639 8.35307 15.1067 8.28353C15.0529 8.21471 14.9823 8.15973 14.9013 8.12344Z" fill="#2FBACE" />
-            <path d="M15.3965 7.8387L15.4469 7.87676C15.4618 7.89119 15.4782 7.90431 15.4931 7.92006C15.52 7.95092 15.542 7.9854 15.5585 8.02241C15.5847 8.09515 15.6069 8.16917 15.6251 8.24416C15.6367 8.2798 15.6532 8.31379 15.6741 8.3452C15.685 8.36095 15.6972 8.37538 15.7095 8.38981C15.7217 8.40425 15.7367 8.41868 15.753 8.43049C15.8134 8.48306 15.8863 8.52046 15.9653 8.5394C16.0482 8.55474 16.1327 8.56003 16.217 8.55515C16.2626 8.5553 16.3081 8.55925 16.3531 8.56696C16.4004 8.57542 16.4462 8.59045 16.4891 8.61157C16.5718 8.65075 16.6436 8.70832 16.6987 8.77953C16.7459 8.8482 16.7766 8.92622 16.7885 9.00784C16.7971 9.08044 16.8232 9.15009 16.8647 9.21123C16.9115 9.27169 16.972 9.32106 17.0415 9.35557C17.1139 9.39122 17.1943 9.40924 17.2756 9.40805C17.3669 9.40479 17.4584 9.41272 17.5477 9.43167C17.6358 9.45679 17.7165 9.50177 17.7831 9.56289C17.7994 9.57732 17.813 9.59438 17.828 9.61013C17.8415 9.62634 17.8537 9.64344 17.8647 9.6613C17.886 9.69657 17.902 9.73456 17.9123 9.77415C17.9237 9.84933 17.9396 9.92381 17.96 9.99722C17.9869 10.0708 18.0339 10.1361 18.096 10.1862L18.0457 10.1481C18.0307 10.1337 18.0144 10.1206 18.0008 10.1048C17.9731 10.0743 17.9506 10.0398 17.9341 10.0025C17.9077 9.92978 17.8854 9.85575 17.8674 9.78071C17.8564 9.74512 17.8403 9.71113 17.8198 9.67967C17.8076 9.66524 17.7967 9.64949 17.7831 9.63506C17.7695 9.62062 17.7559 9.6075 17.7395 9.59438C17.6788 9.54229 17.606 9.50495 17.5273 9.48547C17.4443 9.47075 17.3598 9.46546 17.2756 9.46972C17.2299 9.4701 17.1844 9.46614 17.1395 9.45792C17.0925 9.44885 17.0468 9.43431 17.0034 9.41461C16.9212 9.37418 16.8496 9.31631 16.7939 9.24534C16.7468 9.1767 16.7166 9.09862 16.7055 9.01703C16.6962 8.94457 16.6702 8.87505 16.6293 8.81364C16.5816 8.75349 16.5207 8.70421 16.451 8.6693C16.4159 8.65134 16.3783 8.6381 16.3395 8.62994C16.2991 8.62219 16.2581 8.6178 16.217 8.61682C16.1256 8.62 16.0342 8.61206 15.9449 8.5932C15.8572 8.56803 15.7769 8.52304 15.7108 8.46198C15.6932 8.44755 15.6809 8.43049 15.6646 8.41474C15.6483 8.399 15.6401 8.38063 15.6278 8.36357C15.6068 8.32835 15.5912 8.29033 15.5816 8.25072C15.5612 8.17462 15.5598 8.09982 15.5326 8.02766C15.5058 7.95396 15.4588 7.88866 15.3965 7.8387Z" fill="#2FBACE" />
-            <path d="M15.5435 7.45293L15.5897 7.50541C15.6034 7.52378 15.617 7.54216 15.6292 7.56184C15.6522 7.60275 15.6678 7.64712 15.6755 7.69305C15.6864 7.78013 15.6919 7.86777 15.6918 7.95549C15.696 7.99673 15.7056 8.0373 15.7204 8.07621C15.7288 8.09552 15.7383 8.11435 15.7489 8.13263C15.7599 8.15092 15.7722 8.16845 15.7857 8.18512C15.8408 8.25491 15.9127 8.31075 15.9952 8.34782C16.083 8.38012 16.1743 8.40258 16.2673 8.41474C16.3188 8.42419 16.3693 8.4378 16.4184 8.45542C16.4669 8.47455 16.5126 8.4997 16.5544 8.53022C16.6352 8.59038 16.7 8.66818 16.7436 8.75722C16.7771 8.84069 16.7906 8.93042 16.783 9.01965C16.7743 9.10009 16.7851 9.18137 16.8143 9.25715C16.854 9.33451 16.9123 9.40156 16.9844 9.45267C17.0576 9.50594 17.143 9.54149 17.2334 9.55633C17.3351 9.57028 17.4343 9.5972 17.5286 9.63637C17.6202 9.68035 17.6987 9.74577 17.7572 9.82663C17.7712 9.84597 17.7839 9.86613 17.7953 9.88699C17.806 9.90794 17.8155 9.9294 17.8239 9.95129C17.8375 9.99383 17.8448 10.038 17.8457 10.0825C17.8409 10.1673 17.8409 10.2523 17.8457 10.3371C17.8571 10.4244 17.8949 10.5066 17.9545 10.5733L17.9069 10.5208C17.8947 10.5024 17.8797 10.484 17.8674 10.4643C17.8447 10.4234 17.829 10.379 17.8212 10.3331C17.8111 10.2482 17.8057 10.1628 17.8049 10.0773C17.8013 10.0356 17.7921 9.9946 17.7776 9.95523C17.7681 9.93686 17.76 9.91717 17.7491 9.8988C17.7382 9.88043 17.7259 9.86337 17.711 9.84632C17.6561 9.77622 17.5841 9.72032 17.5014 9.68361C17.4135 9.65186 17.3223 9.62942 17.2293 9.61669C17.1776 9.6082 17.127 9.59457 17.0783 9.57601C17.0294 9.55765 16.9835 9.53245 16.9422 9.50122C16.8607 9.44198 16.7953 9.36453 16.7517 9.27553C16.717 9.1896 16.7035 9.09705 16.7123 9.00522C16.7221 8.92477 16.7114 8.84323 16.681 8.76772C16.6264 8.6512 16.5297 8.55788 16.4088 8.50528C16.3647 8.48882 16.3192 8.47609 16.2728 8.46723C16.1716 8.45326 16.0727 8.42634 15.9789 8.38719C15.8642 8.33186 15.7694 8.24454 15.7068 8.13657C15.6965 8.11561 15.6874 8.09415 15.6796 8.07227C15.6654 8.02982 15.6576 7.98563 15.6564 7.94105C15.6611 7.85582 15.6611 7.77041 15.6564 7.68518C15.6423 7.59903 15.6031 7.51851 15.5435 7.45293Z" fill="#2FBACE" />
-            <path d="M16.1816 7.65894L16.2061 7.70486C16.2061 7.72061 16.2184 7.73636 16.2238 7.75341C16.2293 7.78689 16.2293 7.82098 16.2238 7.85445C16.2127 7.91909 16.1977 7.98305 16.1789 8.04603C16.1748 8.0774 16.1748 8.10913 16.1789 8.1405C16.1789 8.15625 16.1789 8.17199 16.1884 8.18643C16.193 8.202 16.1985 8.21733 16.2048 8.23235C16.2317 8.29366 16.2743 8.34739 16.3286 8.3885C16.3899 8.42743 16.4557 8.45957 16.5245 8.48429C16.5612 8.4993 16.5963 8.51776 16.6293 8.5394C16.6625 8.56142 16.6923 8.58793 16.7177 8.61813C16.7665 8.67301 16.7982 8.74004 16.8093 8.81152C16.8204 8.883 16.8103 8.95605 16.7803 9.02228C16.7566 9.07974 16.7468 9.14168 16.7517 9.20336C16.7622 9.26627 16.7894 9.32549 16.8306 9.37525C16.8746 9.42708 16.9312 9.46765 16.9953 9.49334C17.0684 9.52035 17.1374 9.55653 17.2007 9.60094C17.2621 9.64788 17.3092 9.71001 17.3368 9.78071C17.3435 9.79738 17.3489 9.81448 17.3531 9.83188C17.3567 9.84919 17.359 9.86674 17.3599 9.88437C17.3606 9.91886 17.3565 9.95329 17.3477 9.98672C17.3217 10.048 17.2995 10.1106 17.281 10.1744C17.2707 10.2397 17.2811 10.3065 17.3109 10.3659L17.2865 10.32C17.2865 10.3043 17.2742 10.2885 17.2701 10.2715C17.2614 10.2389 17.2586 10.2052 17.262 10.1717C17.262 10.1035 17.296 10.0405 17.3069 9.97885C17.3123 9.94757 17.3123 9.91565 17.3069 9.88437C17.3069 9.86862 17.3069 9.85288 17.2973 9.83844C17.2933 9.82269 17.2879 9.80732 17.281 9.79252C17.2549 9.73151 17.2127 9.67814 17.1586 9.63768C17.0976 9.598 17.0318 9.56582 16.9626 9.54189C16.926 9.52585 16.8905 9.50744 16.8565 9.48678C16.8237 9.46425 16.7943 9.4373 16.7694 9.40674C16.7204 9.34995 16.688 9.28154 16.6755 9.2086C16.6678 9.13861 16.6781 9.06786 16.7055 9.00259C16.7307 8.94553 16.7405 8.88323 16.734 8.82152C16.7232 8.75934 16.6966 8.70072 16.6565 8.65093C16.6359 8.62499 16.6117 8.60202 16.5844 8.5827C16.5547 8.56306 16.5233 8.54594 16.4905 8.53153C16.4175 8.50424 16.3485 8.46808 16.285 8.42393C16.2238 8.37686 16.1768 8.31477 16.149 8.24416C16.149 8.22842 16.1394 8.21005 16.134 8.19299C16.1286 8.17593 16.134 8.15887 16.1258 8.1405C16.1256 8.10638 16.1302 8.07239 16.1394 8.03947C16.1627 7.97795 16.1827 7.91532 16.1993 7.85183C16.2149 7.78736 16.2087 7.71976 16.1816 7.65894Z" fill="#2FBACE" />
-            <path d="M16.6007 7.48442L16.6102 7.53691C16.6102 7.55528 16.6102 7.57365 16.6102 7.59202C16.6071 7.62806 16.5984 7.66345 16.5844 7.69699C16.5503 7.76103 16.5126 7.82323 16.4714 7.88332C16.4544 7.91303 16.4416 7.9448 16.4333 7.97779C16.4333 7.99485 16.4333 8.01191 16.4333 8.02897C16.432 8.04556 16.432 8.06223 16.4333 8.07883C16.4404 8.14795 16.4673 8.21382 16.5109 8.26909C16.5628 8.3254 16.6217 8.37521 16.6864 8.41737C16.7195 8.44249 16.7504 8.47013 16.7789 8.50004C16.8059 8.53153 16.8283 8.5664 16.8456 8.6037C16.8775 8.67561 16.8883 8.75458 16.8769 8.83201C16.8611 8.90673 16.8265 8.97657 16.7762 9.0354C16.7315 9.08701 16.7002 9.1482 16.6851 9.21385C16.6755 9.28008 16.6844 9.34758 16.7109 9.40937C16.7409 9.47439 16.7878 9.53088 16.847 9.57339C16.9152 9.61877 16.9762 9.67359 17.0279 9.7361C17.0752 9.7995 17.1026 9.87464 17.1069 9.9526C17.1069 9.97097 17.1069 9.99065 17.1069 10.009C17.1022 10.0654 17.0825 10.1196 17.0497 10.1665C17.004 10.2231 16.9618 10.2823 16.9232 10.3436C16.8894 10.407 16.8774 10.4792 16.8892 10.5496L16.8796 10.4958C16.8796 10.4775 16.8796 10.4591 16.8796 10.442C16.8814 10.4054 16.8902 10.3694 16.9055 10.3358C16.9386 10.2715 16.9759 10.2092 17.0171 10.1494C17.0343 10.1198 17.0471 10.088 17.0551 10.055C17.0564 10.0384 17.0564 10.0217 17.0551 10.0051C17.0565 9.98806 17.0565 9.97095 17.0551 9.95391C17.0485 9.88491 17.0221 9.81904 16.979 9.76365C16.9268 9.70766 16.8679 9.65788 16.8034 9.61538C16.7697 9.59096 16.7387 9.56326 16.7109 9.53271C16.6832 9.50171 16.6607 9.46673 16.6442 9.42905C16.6119 9.35726 16.6006 9.27829 16.6116 9.20073C16.6276 9.1261 16.6622 9.05631 16.7123 8.99735C16.7573 8.94591 16.7886 8.88465 16.8034 8.81889C16.8126 8.75277 16.8041 8.6855 16.7789 8.62338C16.7653 8.59169 16.747 8.56207 16.7245 8.53546C16.7 8.50794 16.6726 8.48288 16.6429 8.46067C16.5755 8.41311 16.5155 8.35655 16.4646 8.29271C16.4166 8.2297 16.3891 8.15433 16.3857 8.07621C16.3844 8.05742 16.3844 8.03857 16.3857 8.01978C16.3857 8.00141 16.3857 7.98304 16.3966 7.96467C16.4068 7.92877 16.4229 7.89467 16.4442 7.86364C16.485 7.80065 16.5381 7.74816 16.5708 7.68518C16.6016 7.62273 16.6121 7.55272 16.6007 7.48442Z" fill="#2FBACE" />
-            <path d="M17.0756 7.09077V7.159C17.0756 7.18131 17.066 7.20493 17.0592 7.22724C17.045 7.27193 17.0224 7.31373 16.9926 7.35058C16.9341 7.42668 16.8565 7.4818 16.7885 7.55265C16.7569 7.58467 16.7304 7.62091 16.7095 7.66025C16.7 7.67981 16.6918 7.69998 16.6851 7.72061C16.6792 7.74132 16.6747 7.76236 16.6714 7.78359C16.6572 7.86984 16.669 7.9582 16.7055 8.03815C16.7513 8.11817 16.8076 8.19225 16.8728 8.2586C16.9072 8.29934 16.9373 8.34329 16.9626 8.38981C16.9859 8.43525 17.0015 8.48398 17.0089 8.53415C17.0255 8.63167 17.0132 8.73173 16.9735 8.82283C16.9277 8.91194 16.8611 8.98963 16.7789 9.04983C16.7058 9.10184 16.6458 9.16913 16.6034 9.24666C16.5687 9.32682 16.5582 9.41482 16.5731 9.5005C16.588 9.58617 16.6278 9.66603 16.6878 9.73085C16.7587 9.80449 16.8182 9.88756 16.8647 9.97753C16.9025 10.0692 16.9119 10.1695 16.8919 10.2662C16.8865 10.2893 16.8797 10.3121 16.8715 10.3344C16.8629 10.3573 16.8529 10.3797 16.8415 10.4014C16.8153 10.4426 16.7837 10.4805 16.7476 10.5142C16.6709 10.573 16.5973 10.6357 16.5272 10.7018C16.4653 10.7708 16.4269 10.8565 16.417 10.9472V10.879C16.417 10.8567 16.4265 10.8344 16.4333 10.8121C16.4482 10.7649 16.4712 10.7206 16.5014 10.6809C16.5585 10.6047 16.6374 10.5496 16.7041 10.4788C16.7356 10.4465 16.7626 10.4103 16.7844 10.3712C16.7935 10.3516 16.8012 10.3314 16.8075 10.3108C16.8135 10.2901 16.818 10.2691 16.8211 10.2478C16.836 10.1616 16.8241 10.0731 16.7871 9.99328C16.7405 9.91111 16.6828 9.83522 16.6157 9.76759C16.582 9.72637 16.5519 9.68247 16.5259 9.63637C16.5014 9.59106 16.4845 9.54237 16.4755 9.49203C16.4592 9.39408 16.4715 9.29368 16.5109 9.20204C16.5567 9.11293 16.6233 9.03524 16.7055 8.97504C16.779 8.92341 16.8391 8.85603 16.881 8.77821C16.915 8.69993 16.9263 8.61421 16.9136 8.53022C16.9076 8.48787 16.8948 8.44669 16.8755 8.40818C16.8529 8.36862 16.8269 8.33088 16.798 8.29534C16.7273 8.2221 16.6686 8.13892 16.6238 8.04865C16.5847 7.95776 16.5752 7.85752 16.5966 7.76129C16.6009 7.73762 16.6073 7.71435 16.6157 7.69174C16.6247 7.66928 16.6352 7.64737 16.647 7.62613C16.6724 7.58395 16.704 7.5455 16.7408 7.51198C16.813 7.44505 16.8987 7.39388 16.9599 7.32434C17.022 7.25953 17.0623 7.1781 17.0756 7.09077Z" fill="#2FBACE" />
-            <path d="M17.5354 7.3532L17.5137 7.41356C17.5028 7.43193 17.4933 7.45162 17.481 7.46999C17.4565 7.50787 17.4252 7.5412 17.3885 7.5684C17.315 7.62613 17.2252 7.66156 17.1504 7.71142C17.1133 7.73417 17.0794 7.76152 17.0497 7.79278C17.0375 7.80852 17.0239 7.82427 17.013 7.84133C17.0021 7.85902 16.9925 7.87744 16.9844 7.89644C16.9502 7.96957 16.9388 8.05074 16.9517 8.13C16.974 8.2128 17.0078 8.29227 17.0524 8.36619C17.0743 8.40795 17.0912 8.45197 17.1028 8.49741C17.1133 8.54037 17.1165 8.58468 17.1123 8.62863C17.1023 8.72099 17.0656 8.80877 17.0062 8.88188C16.9414 8.95317 16.861 9.00965 16.7708 9.04721C16.6905 9.07947 16.6185 9.12829 16.5599 9.19023C16.5104 9.25313 16.4786 9.32724 16.4674 9.40543C16.4593 9.4861 16.4758 9.56728 16.515 9.63899C16.5606 9.71818 16.5937 9.80356 16.6129 9.89224C16.6241 9.98233 16.6075 10.0736 16.5653 10.1547C16.5558 10.1744 16.5408 10.1927 16.5299 10.2124C16.5157 10.2312 16.5002 10.2492 16.4837 10.2662C16.4492 10.2985 16.4108 10.3267 16.3694 10.3502C16.2864 10.3961 16.1952 10.425 16.1204 10.4814C16.0466 10.5304 15.9901 10.6 15.9585 10.6809L15.9802 10.6205C15.9893 10.6006 15.9998 10.5813 16.0115 10.5628C16.0372 10.5257 16.0684 10.4926 16.1041 10.4643C16.1775 10.4066 16.2673 10.3712 16.3422 10.3226C16.3793 10.2994 16.4131 10.2716 16.4429 10.24C16.4563 10.2247 16.4686 10.2084 16.4796 10.1914C16.4902 10.174 16.4997 10.1561 16.5082 10.1376C16.5432 10.0642 16.5545 9.98242 16.5408 9.90274C16.5193 9.82033 16.4859 9.74126 16.4415 9.66786C16.4196 9.62596 16.4023 9.58199 16.3898 9.53665C16.3792 9.49103 16.376 9.44412 16.3803 9.39756C16.3911 9.30539 16.4278 9.21783 16.4864 9.14431C16.5512 9.07301 16.6316 9.01653 16.7218 8.97898C16.8024 8.94664 16.8744 8.89732 16.9327 8.83464C16.9835 8.77247 17.0155 8.69801 17.0252 8.61944C17.032 8.58079 17.032 8.54131 17.0252 8.50266C17.0147 8.46244 17.0002 8.42332 16.9817 8.38588C16.9362 8.30672 16.9036 8.22131 16.8851 8.13263C16.8731 8.04252 16.8897 7.95103 16.9327 7.8702C16.9426 7.84954 16.9545 7.82977 16.9681 7.81115C16.9819 7.79276 16.9969 7.77523 17.013 7.75866C17.0468 7.7256 17.0852 7.69732 17.1273 7.67468C17.2116 7.62745 17.3028 7.59858 17.3763 7.54347C17.4472 7.49616 17.5025 7.43009 17.5354 7.3532Z" fill="#2FBACE" />
-            <path d="M15.7653 10.5431C15.7955 10.5209 15.8233 10.4958 15.8483 10.4683C15.8742 10.4407 15.9029 10.4156 15.934 10.3935C16.0456 10.3554 16.1626 10.3266 16.266 10.2806C16.3158 10.2601 16.3618 10.2317 16.402 10.1967C16.4403 10.1629 16.4701 10.1212 16.4891 10.0746C16.5211 9.97819 16.5211 9.87453 16.4891 9.77808L16.4483 9.62194C16.4336 9.56795 16.4258 9.51242 16.4252 9.4566C16.4227 9.39192 16.4362 9.32761 16.4646 9.26896C16.49 9.20909 16.5275 9.15467 16.5748 9.10888C16.6662 9.02198 16.766 8.94377 16.8728 8.87531C16.915 8.83743 16.9504 8.7931 16.9776 8.7441C17.0041 8.69368 17.0207 8.63896 17.0266 8.5827V8.59582C16.9953 8.49216 16.9517 8.3885 16.9245 8.27959C16.9096 8.22471 16.9014 8.16835 16.9 8.11163C16.8999 8.05404 16.9114 7.99697 16.9341 7.94368C16.9464 7.91734 16.9614 7.89228 16.979 7.86888C16.9994 7.84658 17.0184 7.82558 17.0402 7.80459C17.0875 7.76621 17.1414 7.73603 17.1994 7.71536C17.3151 7.67527 17.4332 7.64196 17.5531 7.61564C17.5851 7.59551 17.6152 7.57268 17.6429 7.5474C17.6716 7.52293 17.702 7.50056 17.7341 7.48048C17.7045 7.50335 17.6768 7.52835 17.6511 7.55528C17.6253 7.58253 17.5971 7.60756 17.5667 7.63007C17.4512 7.66242 17.3377 7.70097 17.2266 7.74554C17.1762 7.76683 17.1301 7.7966 17.0905 7.83346C17.0722 7.8522 17.0549 7.87191 17.0388 7.8925C17.0246 7.91355 17.0128 7.93598 17.0034 7.95942C16.9695 8.05736 16.9695 8.16328 17.0034 8.26122C17.032 8.36357 17.0769 8.46592 17.1123 8.57352V8.58664C17.1066 8.65093 17.0886 8.71365 17.0592 8.77165C17.0292 8.83049 16.9887 8.88379 16.9395 8.92911C16.8341 8.99616 16.7345 9.07119 16.6415 9.15349C16.6006 9.19062 16.5678 9.23522 16.5449 9.28471C16.5249 9.33577 16.5161 9.39028 16.5191 9.44479C16.526 9.5487 16.5443 9.65161 16.5735 9.75184C16.5853 9.80579 16.5917 9.86072 16.5925 9.91586C16.5935 9.97225 16.5819 10.0282 16.5585 10.0799C16.5332 10.1304 16.4971 10.1752 16.4524 10.2111C16.4045 10.247 16.3513 10.2758 16.2946 10.2964C16.1816 10.3384 16.0619 10.3594 15.949 10.3896C15.9173 10.4096 15.8873 10.432 15.8592 10.4565C15.8309 10.4882 15.7994 10.5172 15.7653 10.5431Z" fill="#FF2316" />
-            <path d="M15.6401 10.0576C15.6686 10.0443 15.6963 10.0294 15.7231 10.013C15.7493 9.99533 15.7771 9.97996 15.8061 9.96704C15.9033 9.95965 16 9.94739 16.0959 9.93029C16.1435 9.92335 16.1895 9.90871 16.232 9.88699C16.2692 9.86589 16.3015 9.83777 16.3272 9.80433C16.3722 9.73493 16.3936 9.65375 16.3884 9.57207V9.44086C16.3884 9.39672 16.3934 9.35271 16.4034 9.30964C16.4127 9.25759 16.4341 9.20825 16.466 9.1653C16.4997 9.12252 16.5412 9.08606 16.5885 9.05771C16.6828 9.00505 16.7821 8.96112 16.8851 8.92649C16.9281 8.90255 16.9667 8.87196 16.9994 8.83595C17.0298 8.80055 17.0541 8.76068 17.0715 8.71786V8.72835C17.0715 8.64175 17.0565 8.55383 17.0565 8.46592C17.0569 8.42173 17.0624 8.37772 17.0728 8.3347C17.0831 8.28694 17.1045 8.24206 17.1354 8.20349C17.1502 8.18423 17.1676 8.16702 17.1871 8.15231C17.2069 8.13726 17.2278 8.12366 17.2497 8.11163C17.2952 8.08854 17.3444 8.07301 17.3953 8.06571C17.4974 8.04996 17.5994 8.05259 17.6987 8.04603C17.7286 8.03478 17.7577 8.02163 17.7858 8.00666C17.8138 7.99187 17.8429 7.97915 17.8729 7.96861C17.8448 7.98191 17.8175 7.99681 17.7912 8.01322C17.7645 8.03084 17.7363 8.0462 17.7069 8.05915C17.6076 8.07096 17.5055 8.07489 17.4116 8.09458C17.3641 8.10363 17.3182 8.11955 17.2756 8.14181C17.2565 8.15362 17.2388 8.16806 17.2211 8.18118C17.2053 8.19539 17.1912 8.21123 17.179 8.22842C17.1328 8.29868 17.1122 8.38179 17.1205 8.46461C17.1205 8.5499 17.1368 8.63781 17.1422 8.72704V8.73754C17.1236 8.78572 17.0959 8.83019 17.0606 8.86875C17.0219 8.91109 16.9758 8.94661 16.9245 8.97373C16.8263 9.01112 16.7308 9.05496 16.6388 9.10494C16.5983 9.12953 16.5624 9.16059 16.5327 9.1968C16.5058 9.2333 16.4872 9.27489 16.4782 9.31883C16.4622 9.40054 16.4553 9.48369 16.4578 9.56683C16.4556 9.61103 16.4492 9.65495 16.4388 9.69804C16.4285 9.74316 16.4076 9.78538 16.3776 9.82139C16.3469 9.85902 16.3078 9.88952 16.2633 9.91061C16.2178 9.93227 16.1692 9.94689 16.119 9.95391C16.0197 9.96835 15.9204 9.96572 15.8224 9.97228C15.7928 9.98327 15.7642 9.99643 15.7367 10.0116C15.7059 10.0295 15.6736 10.0449 15.6401 10.0576Z" fill="#FF2316" />
-            <path d="M14.9462 9.99065C14.9873 9.98251 15.0278 9.97155 15.0673 9.95785C15.1077 9.94416 15.1491 9.93319 15.1911 9.92505C15.319 9.93948 15.4469 9.96179 15.5721 9.96441C15.633 9.96644 15.6939 9.96026 15.753 9.94604C15.8083 9.93078 15.8601 9.90539 15.9054 9.87125C15.9871 9.79906 16.0444 9.7049 16.0701 9.60094L16.119 9.44217C16.1353 9.38567 16.1591 9.33142 16.1898 9.28077C16.2221 9.21825 16.2686 9.16353 16.3259 9.12069C16.3847 9.07678 16.451 9.04303 16.5218 9.02096C16.6617 8.98265 16.8045 8.95544 16.949 8.93961C17.0125 8.92239 17.0727 8.89536 17.1273 8.85957C17.18 8.82358 17.226 8.77921 17.2633 8.72835V8.74016C17.2919 8.63388 17.3082 8.52103 17.3436 8.4095C17.3608 8.35288 17.3845 8.2983 17.4144 8.24679C17.4472 8.1932 17.4906 8.14639 17.5423 8.10901C17.5687 8.08995 17.597 8.07326 17.6266 8.05915C17.6579 8.04734 17.6892 8.03553 17.7219 8.02634C17.7887 8.01164 17.8576 8.00677 17.926 8.01191C18.062 8.0211 18.1886 8.0539 18.3178 8.07489C18.36 8.07489 18.4008 8.06046 18.443 8.05127C18.484 8.04013 18.5258 8.03223 18.5682 8.02766C18.527 8.0358 18.4866 8.04676 18.4471 8.06046C18.4068 8.07476 18.3654 8.08573 18.3233 8.09326C18.1872 8.08014 18.062 8.05652 17.9341 8.05521C17.8721 8.05273 17.8101 8.05981 17.7504 8.07621C17.7222 8.08578 17.6945 8.09673 17.6674 8.10901C17.6424 8.12234 17.6183 8.13724 17.5953 8.15362C17.5117 8.22681 17.4545 8.3238 17.432 8.43049C17.4007 8.53678 17.3871 8.64831 17.3599 8.76116V8.76772C17.3111 8.83157 17.2502 8.8859 17.1803 8.9278C17.1178 8.96898 17.0489 9.00042 16.9762 9.02096H16.9694C16.8316 9.03729 16.6952 9.06316 16.5612 9.09838C16.5 9.11673 16.4424 9.14514 16.3912 9.18236C16.3444 9.21832 16.3055 9.26293 16.2769 9.31358C16.2265 9.41049 16.1864 9.51204 16.1571 9.61669C16.1383 9.67104 16.1137 9.72337 16.0837 9.77284C16.0525 9.82456 16.0103 9.86929 15.9598 9.90405C15.9064 9.94108 15.8458 9.96741 15.7816 9.98147C15.7158 9.99454 15.6485 9.99851 15.5816 9.99328C15.4455 9.98278 15.3244 9.95129 15.1979 9.93161C15.1561 9.93644 15.1147 9.94433 15.0741 9.95523C15.0323 9.96986 14.9896 9.98169 14.9462 9.99065Z" fill="#FF2316" />
-            <path d="M14.8455 9.44217C14.8849 9.4448 14.9244 9.4448 14.9639 9.44217C15.0042 9.43955 15.0446 9.43955 15.085 9.44217C15.1952 9.48285 15.3013 9.53271 15.4129 9.56289C15.4666 9.57927 15.5225 9.5881 15.5789 9.58913C15.6328 9.58854 15.6863 9.57877 15.7367 9.56026C15.8305 9.51662 15.91 9.44899 15.9666 9.36475L16.0578 9.23354C16.0901 9.18816 16.1281 9.14674 16.1707 9.11019C16.2185 9.06353 16.2754 9.0265 16.3381 9.00128C16.4039 8.97735 16.4733 8.96363 16.5436 8.96061C16.6801 8.95957 16.8166 8.96878 16.9517 8.98816C17.0136 8.9884 17.0752 8.9791 17.1341 8.96061C17.1915 8.9431 17.2457 8.91697 17.2946 8.88319L17.2851 8.89237C17.3436 8.80708 17.3939 8.71392 17.4579 8.62994C17.4911 8.58486 17.5295 8.54349 17.5722 8.5066C17.617 8.46726 17.6692 8.43653 17.7259 8.41606C17.7556 8.40617 17.7861 8.3987 17.8171 8.39375C17.8484 8.39375 17.8797 8.38719 17.9123 8.38719C17.9754 8.38978 18.0378 8.40128 18.0974 8.42131C18.2097 8.46385 18.3196 8.51203 18.4267 8.56564C18.4654 8.57151 18.5045 8.57415 18.5437 8.57352C18.5837 8.57457 18.6237 8.57764 18.6634 8.5827C18.624 8.58008 18.5844 8.58008 18.545 8.5827C18.5052 8.58598 18.4651 8.58598 18.4253 8.5827C18.3124 8.54071 18.2035 8.48954 18.0906 8.45936C18.0355 8.44326 17.9781 8.4353 17.9205 8.43574C17.8919 8.43574 17.8647 8.43574 17.8361 8.44492C17.8097 8.45036 17.7838 8.45782 17.7586 8.46723C17.6575 8.50705 17.5719 8.57633 17.5137 8.66537C17.4538 8.75066 17.4062 8.84382 17.349 8.9278C17.2937 8.96761 17.2322 8.99862 17.1667 9.01965C17.0989 9.04053 17.0283 9.05159 16.9572 9.05246C16.8283 9.03485 16.6982 9.02608 16.568 9.02621C16.5076 9.0276 16.4478 9.03868 16.3912 9.05902C16.3377 9.07978 16.2888 9.11005 16.2469 9.14824C16.1722 9.21886 16.1052 9.29669 16.0469 9.3805C16.0133 9.42369 15.9755 9.46367 15.934 9.49991C15.8897 9.53867 15.8379 9.56856 15.7816 9.58782C15.7233 9.60842 15.6613 9.61735 15.5993 9.61406C15.5375 9.61039 15.4766 9.59847 15.4183 9.57863C15.304 9.53927 15.202 9.48416 15.0958 9.43692C15.0529 9.43976 15.0102 9.44502 14.9679 9.45267C14.9269 9.45259 14.8859 9.44908 14.8455 9.44217Z" fill="#FF2316" />
-            <path d="M15.1271 9.1115C15.159 9.11823 15.1913 9.12305 15.2237 9.12594C15.2568 9.12862 15.2895 9.13389 15.3217 9.14168C15.4033 9.18892 15.4823 9.24141 15.5666 9.28077C15.6098 9.30067 15.6556 9.31479 15.7027 9.32276C15.7476 9.33099 15.7938 9.33099 15.8387 9.32276C15.9228 9.29927 15.9983 9.25333 16.0564 9.19023L16.153 9.10101C16.1873 9.06934 16.2252 9.0416 16.266 9.01834C16.3125 8.98672 16.3653 8.96482 16.4211 8.95404C16.4783 8.94491 16.5367 8.94491 16.5939 8.95404C16.7042 8.9703 16.813 8.99531 16.9191 9.02884C16.9701 9.03605 17.0219 9.03605 17.0728 9.02884C17.1229 9.02208 17.1715 9.00791 17.2171 8.98685H17.2075C17.2688 8.92649 17.3259 8.85563 17.3939 8.79658C17.4283 8.7643 17.4661 8.7357 17.5069 8.71129C17.549 8.68708 17.595 8.66979 17.6429 8.66012C17.6687 8.65753 17.6947 8.65753 17.7205 8.66012C17.7463 8.66012 17.7722 8.66012 17.7994 8.66012C17.8498 8.67025 17.8984 8.68749 17.9436 8.71129C18.0276 8.75966 18.1089 8.81222 18.1872 8.86875C18.2184 8.87819 18.2502 8.88564 18.2824 8.89106C18.3148 8.89601 18.3466 8.90347 18.3777 8.91337C18.3459 8.906 18.3136 8.90117 18.2811 8.89893C18.2481 8.89561 18.2154 8.89035 18.1831 8.88319C18.0988 8.83464 18.0198 8.78084 17.9328 8.74147C17.8897 8.72152 17.8439 8.70781 17.7967 8.7008C17.7749 8.7008 17.7504 8.7008 17.7273 8.7008C17.7049 8.70257 17.6826 8.70564 17.6606 8.70998C17.5746 8.73097 17.4972 8.77674 17.4388 8.8412C17.3749 8.90156 17.3218 8.97241 17.2592 9.0354C17.2078 9.06002 17.1527 9.07684 17.096 9.08526C17.0373 9.09313 16.9778 9.09313 16.9191 9.08526C16.8171 9.05465 16.713 9.03097 16.6075 9.0144C16.558 9.00583 16.5073 9.00583 16.4578 9.0144C16.4094 9.02315 16.3631 9.04098 16.3218 9.06689C16.2494 9.11375 16.1819 9.16737 16.1204 9.22697C16.0851 9.25657 16.0473 9.28334 16.0075 9.30702C15.9658 9.33204 15.9196 9.34939 15.8714 9.35819C15.8206 9.36736 15.7685 9.36736 15.7176 9.35819C15.6701 9.3471 15.6243 9.32989 15.5816 9.30702C15.4995 9.25937 15.42 9.20769 15.3435 9.15218C15.3122 9.14318 15.2804 9.13617 15.2482 9.13119C15.2074 9.1276 15.1669 9.12102 15.1271 9.1115Z" fill="#FF2316" />
-            <path d="M14.8319 8.69292C14.8669 8.70926 14.9027 8.72371 14.9394 8.73622C14.9763 8.74793 15.0123 8.76241 15.0469 8.77953C15.1271 8.85563 15.1992 8.93699 15.2836 9.00259C15.3245 9.03649 15.3703 9.06435 15.4197 9.08526C15.4695 9.10462 15.5225 9.11528 15.5761 9.11675C15.6821 9.11378 15.7856 9.08488 15.8768 9.03277L16.0211 8.95798C16.0719 8.9304 16.1257 8.90839 16.1816 8.89237C16.2469 8.8694 16.3162 8.85871 16.3857 8.86088C16.4555 8.86319 16.5245 8.87649 16.5898 8.90024C16.7115 8.94712 16.8289 9.00374 16.9409 9.06951C16.9977 9.09235 17.0578 9.10694 17.1191 9.11282C17.18 9.11767 17.2414 9.11458 17.3014 9.10363H17.2878C17.3817 9.05114 17.4715 8.98816 17.5722 8.93567C17.6242 8.90956 17.6784 8.88762 17.7341 8.87006C17.7938 8.85383 17.8558 8.84674 17.9178 8.84907C17.9489 8.85142 17.9798 8.8558 18.0103 8.86219C18.0402 8.87138 18.0702 8.87925 18.0988 8.89106C18.154 8.91565 18.205 8.94841 18.2498 8.98816C18.3302 9.0649 18.4061 9.14594 18.477 9.23091C18.5099 9.24918 18.544 9.2654 18.5791 9.27946C18.6146 9.29441 18.6491 9.3115 18.6825 9.33063C18.6478 9.31395 18.6119 9.29991 18.575 9.28865C18.5381 9.2767 18.5022 9.26223 18.4675 9.24534C18.3872 9.16924 18.3124 9.08657 18.2253 9.01834C18.1844 8.98479 18.1386 8.95737 18.0892 8.93699C18.0634 8.9278 18.0362 8.92124 18.0103 8.91468C17.9834 8.90968 17.9561 8.90661 17.9287 8.90549C17.8205 8.90789 17.7152 8.93874 17.6239 8.99472C17.53 9.0459 17.4429 9.10888 17.3518 9.16661C17.2834 9.18085 17.2133 9.18527 17.1436 9.17974C17.0735 9.17305 17.0048 9.1567 16.9395 9.13119C16.8313 9.06832 16.7185 9.01307 16.6021 8.96585C16.5456 8.94593 16.4865 8.93399 16.4265 8.93042C16.3679 8.92949 16.3095 8.93792 16.2537 8.95536C16.152 8.98971 16.054 9.03366 15.9612 9.08657C15.9097 9.1104 15.8565 9.13058 15.802 9.14693C15.6844 9.18252 15.5573 9.17457 15.4455 9.12462C15.3919 9.09998 15.3428 9.06719 15.2999 9.02753C15.2052 8.95559 15.1148 8.87848 15.0292 8.79658C14.9963 8.77789 14.9622 8.76123 14.9271 8.74672C14.894 8.7311 14.8622 8.71311 14.8319 8.69292Z" fill="#FF2316" />
-            <path d="M14.8019 8.1838C14.8334 8.21003 14.8671 8.23373 14.9026 8.25466C14.9379 8.27693 14.9715 8.30148 15.0033 8.32814C15.0632 8.42787 15.1122 8.53022 15.1802 8.62207C15.2128 8.66811 15.2527 8.70886 15.2986 8.74279C15.3456 8.77631 15.3987 8.80122 15.455 8.81627C15.5711 8.84162 15.6919 8.83801 15.8061 8.80577L15.9857 8.76509C16.0496 8.75071 16.1147 8.74192 16.1803 8.73885C16.257 8.73189 16.3344 8.73811 16.4088 8.75722C16.4816 8.77823 16.5504 8.81011 16.6129 8.8517C16.7269 8.92992 16.8335 9.01774 16.9313 9.11413C16.9853 9.15269 17.0449 9.18321 17.1082 9.20467C17.1717 9.22671 17.2381 9.23954 17.3055 9.24272H17.281C17.3994 9.21254 17.5178 9.17055 17.6429 9.143C17.7071 9.12766 17.7728 9.11886 17.8389 9.11675C17.9079 9.11395 17.9769 9.12238 18.043 9.14168C18.0757 9.15188 18.1076 9.1646 18.1382 9.17974C18.1669 9.1957 18.1946 9.21322 18.2212 9.23222C18.2741 9.27115 18.3201 9.31813 18.3573 9.37131C18.4134 9.47459 18.4629 9.58107 18.5056 9.69017C18.5341 9.71798 18.5645 9.74385 18.5967 9.76759C18.6298 9.79183 18.6607 9.8186 18.6893 9.84763C18.6582 9.82109 18.625 9.79696 18.5899 9.77546C18.5541 9.75403 18.5204 9.72943 18.4893 9.70198C18.428 9.60094 18.3777 9.49597 18.3083 9.40412C18.2742 9.3572 18.2324 9.31597 18.1845 9.28209C18.16 9.26634 18.1328 9.25191 18.1069 9.23747C18.08 9.22591 18.0523 9.21626 18.0239 9.2086C17.9056 9.18208 17.782 9.18661 17.6661 9.22173C17.5477 9.25059 17.4334 9.2939 17.3109 9.32801H17.3014C17.2244 9.32554 17.1483 9.31183 17.0756 9.28733C17.0033 9.26257 16.9351 9.22811 16.8728 9.18499C16.7785 9.09047 16.6771 9.00275 16.5694 8.92255C16.5152 8.88721 16.4561 8.85939 16.3939 8.83989C16.3309 8.82394 16.2657 8.81774 16.2007 8.82152C16.0804 8.83116 15.9611 8.85047 15.8442 8.87925C15.781 8.892 15.7168 8.89947 15.6523 8.90156C15.5852 8.9048 15.518 8.89681 15.4537 8.87794C15.3904 8.85769 15.3314 8.82657 15.2795 8.78609C15.2313 8.74897 15.1895 8.70467 15.1557 8.65487C15.0983 8.55575 15.0465 8.45369 15.0006 8.34914C14.9722 8.32085 14.9417 8.29455 14.9094 8.27041C14.8717 8.24379 14.8358 8.21487 14.8019 8.1838Z" fill="#FF2316" />
-            <path d="M15.104 7.93056C15.1276 7.95997 15.1536 7.9876 15.1816 8.01322C15.2099 8.03899 15.2358 8.06709 15.2591 8.0972C15.2872 8.19829 15.3213 8.29774 15.3612 8.39506C15.3812 8.44316 15.4098 8.48755 15.4455 8.52628C15.4842 8.5655 15.5304 8.59714 15.5816 8.61944C15.6864 8.66165 15.8005 8.67832 15.9136 8.66799L16.0905 8.66012C16.1532 8.65693 16.216 8.65956 16.2782 8.66799C16.3514 8.67393 16.4229 8.69261 16.4891 8.7231C16.5526 8.75374 16.6105 8.79406 16.6606 8.84251C16.7496 8.93501 16.8298 9.03515 16.9 9.14168C16.9426 9.18556 16.9917 9.22314 17.0456 9.25322C17.101 9.28357 17.1605 9.30611 17.2225 9.32014H17.2075C17.3259 9.31227 17.4443 9.29258 17.5695 9.28865C17.6326 9.28477 17.696 9.28741 17.7586 9.29652C17.8231 9.30531 17.8856 9.32438 17.9436 9.35294C17.9722 9.36676 17.9991 9.38348 18.0239 9.40281C18.047 9.4238 18.0715 9.44348 18.0919 9.46579C18.1305 9.50989 18.1608 9.56005 18.1817 9.61406C18.2158 9.71683 18.2431 9.82157 18.2634 9.92767C18.2841 9.95825 18.3073 9.98721 18.3328 10.0143C18.3578 10.0421 18.381 10.0715 18.4022 10.1022C18.3785 10.0732 18.3526 10.046 18.3246 10.0208C18.2962 9.99478 18.2698 9.96671 18.2457 9.93686C18.2174 9.83353 18.1829 9.73189 18.1423 9.63243C18.1202 9.58416 18.0898 9.53981 18.0525 9.50122C18.0326 9.48228 18.0117 9.46432 17.9899 9.44742C17.967 9.43213 17.9429 9.41853 17.9178 9.40674C17.8113 9.36373 17.6951 9.34792 17.5804 9.36082C17.462 9.36082 17.345 9.38837 17.2225 9.40018H17.2075C17.1366 9.38471 17.0683 9.35952 17.0048 9.32539C16.9424 9.28954 16.8856 9.24536 16.8361 9.19417C16.7673 9.09273 16.6905 8.99665 16.6061 8.90681C16.5641 8.86449 16.5149 8.82943 16.4606 8.80315C16.4043 8.77729 16.3441 8.76045 16.2823 8.75328C16.1669 8.74335 16.0508 8.74335 15.9353 8.75328C15.8734 8.75386 15.8116 8.74992 15.7503 8.74147C15.6871 8.73347 15.6258 8.71481 15.5693 8.68636C15.5142 8.65802 15.4652 8.61971 15.4251 8.57352C15.3887 8.52937 15.3602 8.47971 15.3408 8.42655C15.3081 8.32583 15.2817 8.22329 15.2618 8.11951C15.2403 8.08946 15.2172 8.06055 15.1924 8.0329C15.1587 8.00246 15.129 7.96808 15.104 7.93056Z" fill="#FF2316" />
-            <path d="M15.3258 7.45424C15.3416 7.49095 15.3603 7.52647 15.3816 7.56053C15.4046 7.59473 15.4238 7.63125 15.4387 7.66944C15.4362 7.78326 15.4398 7.89714 15.4496 8.0106C15.4555 8.06643 15.4721 8.12072 15.4986 8.17068C15.525 8.22159 15.5627 8.26636 15.6088 8.3019C15.7106 8.37221 15.8289 8.41728 15.953 8.43312L16.1476 8.46986C16.2165 8.48149 16.284 8.49996 16.349 8.52497C16.5003 8.57188 16.6274 8.6728 16.7041 8.80708C16.77 8.92932 16.8229 9.0577 16.8619 9.19023C16.893 9.24771 16.9334 9.29999 16.9817 9.34507C17.0308 9.38794 17.0857 9.42415 17.1449 9.45267H17.13C17.266 9.47366 17.4021 9.48285 17.5382 9.50909C17.6079 9.52144 17.6763 9.54034 17.7423 9.56551C17.8085 9.59137 17.8701 9.62725 17.9246 9.6718C17.9503 9.69433 17.9739 9.71891 17.9953 9.74528C18.0142 9.77233 18.0314 9.80037 18.047 9.82926C18.0739 9.88598 18.09 9.94685 18.0947 10.009C18.0975 10.1265 18.092 10.244 18.0783 10.3607C18.0908 10.3981 18.1063 10.4345 18.1246 10.4696C18.1427 10.5051 18.1577 10.5419 18.1695 10.5798C18.1534 10.5432 18.1347 10.5077 18.1137 10.4735C18.0907 10.4393 18.0715 10.4028 18.0566 10.3646C18.0605 10.2486 18.0569 10.1325 18.0457 10.0169C18.0384 9.96017 18.0209 9.9051 17.994 9.85419C17.979 9.82926 17.96 9.80564 17.9423 9.78202C17.9227 9.75986 17.9013 9.73923 17.8783 9.72035C17.7751 9.648 17.6541 9.60279 17.5273 9.58913C17.3912 9.56683 17.262 9.55895 17.1191 9.54058H17.1109C17.0391 9.50674 16.973 9.46249 16.915 9.40937C16.8605 9.35739 16.8146 9.29764 16.7789 9.23222C16.7396 9.10773 16.6886 8.98694 16.6265 8.87138C16.5946 8.81533 16.5527 8.76513 16.5027 8.7231C16.4502 8.68185 16.3909 8.64947 16.3272 8.62732C16.2052 8.58533 16.0795 8.55414 15.9517 8.53415C15.8846 8.51868 15.8187 8.49896 15.7544 8.4751C15.6893 8.45064 15.629 8.41557 15.5761 8.37144C15.5259 8.32788 15.4856 8.2747 15.4578 8.2153C15.4335 8.15942 15.4192 8.09996 15.4156 8.03947C15.4141 7.92423 15.4196 7.809 15.4319 7.69437C15.4194 7.65697 15.404 7.62055 15.3857 7.58546C15.3612 7.54377 15.3411 7.49981 15.3258 7.45424Z" fill="#FF2316" />
-            <path d="M15.8728 7.31777C15.8771 7.35425 15.8848 7.39026 15.8959 7.42537C15.9065 7.46097 15.9138 7.49742 15.9177 7.53428C15.8828 7.63534 15.8541 7.7383 15.8319 7.84264C15.8217 7.89334 15.8217 7.94546 15.8319 7.99616C15.8433 8.04701 15.8651 8.09518 15.8959 8.13788C15.9708 8.22381 16.0673 8.28977 16.1762 8.32945L16.3449 8.40425C16.4049 8.42872 16.462 8.45952 16.515 8.4961C16.5782 8.53354 16.6332 8.58258 16.6769 8.64044C16.7168 8.69581 16.7463 8.75756 16.764 8.82283C16.7911 8.9449 16.8052 9.06933 16.8062 9.19417C16.8184 9.25213 16.841 9.30762 16.8728 9.35819C16.9068 9.41014 16.9485 9.457 16.9966 9.49728L16.9844 9.48941C17.0987 9.53533 17.2211 9.57339 17.3395 9.62062C17.3997 9.64622 17.4571 9.67742 17.511 9.71379C17.564 9.74939 17.6101 9.79382 17.647 9.845C17.6646 9.87023 17.6797 9.89703 17.6919 9.92505C17.7001 9.95391 17.711 9.98147 17.7164 10.0103C17.7273 10.0671 17.7273 10.1254 17.7164 10.1822C17.6861 10.2887 17.6484 10.393 17.6035 10.4945C17.6056 10.5306 17.6097 10.5665 17.6157 10.6021C17.6226 10.6381 17.6267 10.6745 17.628 10.711C17.623 10.6747 17.6153 10.6387 17.6048 10.6034C17.5935 10.568 17.5862 10.5315 17.5831 10.4945C17.6184 10.3917 17.6475 10.287 17.6702 10.1809C17.6804 10.1289 17.6804 10.0755 17.6702 10.0235C17.6634 9.99797 17.6547 9.97297 17.6443 9.94867C17.6324 9.92476 17.6183 9.90192 17.6021 9.88043C17.5272 9.79162 17.4283 9.72443 17.3164 9.68623C17.2021 9.63899 17.081 9.60357 16.9599 9.55502H16.9531C16.8969 9.50999 16.8483 9.45688 16.8089 9.39756C16.7716 9.33935 16.7449 9.27545 16.73 9.2086C16.7274 9.0892 16.7142 8.97023 16.6905 8.85301C16.6765 8.79613 16.6516 8.74227 16.617 8.69424C16.5795 8.64573 16.5335 8.60397 16.481 8.57089C16.3798 8.50727 16.2728 8.4528 16.1612 8.40818C16.1043 8.38067 16.0493 8.34954 15.9966 8.31502C15.943 8.27995 15.8969 8.23543 15.8605 8.1838C15.8262 8.13457 15.8039 8.0785 15.7952 8.01978C15.7881 7.96369 15.7913 7.90683 15.8047 7.85183C15.8354 7.74823 15.8726 7.64656 15.9163 7.5474C15.9155 7.51126 15.9114 7.47525 15.9041 7.43981C15.8901 7.40005 15.8796 7.35923 15.8728 7.31777Z" fill="#FF2316" />
-            <path d="M16.4129 7.42931C16.408 7.46057 16.4057 7.49217 16.4061 7.52378C16.4061 7.55588 16.4034 7.58791 16.398 7.61957C16.3435 7.69778 16.2935 7.77882 16.2483 7.86232C16.2273 7.90341 16.2144 7.9479 16.2102 7.99354C16.2078 8.03827 16.2152 8.08299 16.232 8.12476C16.275 8.21132 16.3418 8.28484 16.4252 8.33733L16.5531 8.4318C16.5977 8.4634 16.6388 8.49946 16.6755 8.5394C16.7193 8.58314 16.7544 8.6343 16.7789 8.6903C16.8001 8.74481 16.8102 8.8027 16.8089 8.86088C16.802 8.96868 16.7838 9.07551 16.7544 9.17974C16.7515 9.23021 16.7575 9.28079 16.7721 9.32932C16.7884 9.37948 16.8127 9.42686 16.8443 9.46972L16.832 9.45267C16.9204 9.51171 17.0157 9.56683 17.1041 9.63243C17.1489 9.66488 17.1903 9.70134 17.2279 9.74134C17.2646 9.77963 17.2937 9.82411 17.3137 9.87256C17.3232 9.89708 17.3296 9.92264 17.3327 9.94867C17.3348 9.974 17.3348 9.99944 17.3327 10.0248C17.3265 10.075 17.3103 10.1235 17.2851 10.1678C17.2322 10.2512 17.174 10.3314 17.1109 10.4079C17.1034 10.4386 17.0979 10.4697 17.0946 10.5011C17.0919 10.5325 17.0865 10.5637 17.0783 10.5942C17.081 10.5628 17.081 10.5312 17.0783 10.4998C17.0769 10.4676 17.0797 10.4355 17.0864 10.404C17.142 10.3241 17.1929 10.2413 17.2388 10.156C17.2594 10.1149 17.2715 10.0703 17.2742 10.0248C17.2762 10.0025 17.2762 9.98011 17.2742 9.95785C17.2699 9.93547 17.2636 9.91351 17.2552 9.89224C17.2113 9.80356 17.1424 9.72855 17.0565 9.67573C16.9708 9.61538 16.8742 9.5642 16.7844 9.50384C16.7473 9.45584 16.7187 9.40217 16.7 9.34507C16.6828 9.28891 16.6754 9.23037 16.6783 9.17186C16.705 9.07116 16.7227 8.96843 16.7313 8.86482C16.7335 8.81517 16.7252 8.76561 16.7068 8.71917C16.6876 8.67038 16.6584 8.62578 16.6211 8.58795C16.5488 8.5162 16.47 8.45078 16.3857 8.39244C16.3433 8.3592 16.3038 8.32279 16.2673 8.28353C16.2305 8.24559 16.2018 8.20099 16.183 8.15231C16.1661 8.10423 16.1605 8.05314 16.1667 8.00272C16.1742 7.95372 16.1913 7.90653 16.217 7.86364C16.2663 7.77795 16.3208 7.69514 16.3803 7.61564C16.3891 7.58523 16.3946 7.554 16.3966 7.52247C16.3993 7.49103 16.4048 7.45986 16.4129 7.42931Z" fill="#FF2316" />
-            <path d="M16.8606 7.6222C16.8484 7.6476 16.8384 7.67394 16.8306 7.70093C16.8228 7.72765 16.8123 7.75357 16.7993 7.77834C16.7331 7.83315 16.6699 7.89141 16.6102 7.95286C16.582 7.98227 16.5599 8.01655 16.5449 8.0539C16.5311 8.08993 16.526 8.12852 16.5299 8.16675C16.5454 8.24718 16.5841 8.32177 16.6415 8.38194L16.7272 8.48429C16.7574 8.51973 16.7838 8.55797 16.8062 8.59845C16.8338 8.64344 16.8514 8.69349 16.8579 8.74541C16.8623 8.79505 16.8563 8.84504 16.8402 8.89237C16.8073 8.98397 16.7644 9.07194 16.7123 9.15481C16.6969 9.19696 16.6891 9.24133 16.6891 9.28602C16.6908 9.33089 16.7 9.3752 16.7164 9.41724V9.40805C16.7762 9.47497 16.8524 9.53927 16.9055 9.61013C16.9361 9.6457 16.9626 9.68443 16.9844 9.7256C17.006 9.76656 17.0194 9.81107 17.0239 9.85681C17.0259 9.87952 17.0259 9.90234 17.0239 9.92505C17.0188 9.94638 17.0124 9.96741 17.0048 9.98803C16.9868 10.0291 16.961 10.0665 16.9286 10.0983C16.8618 10.1586 16.7914 10.2152 16.7177 10.2675C16.7036 10.2915 16.6913 10.3165 16.681 10.3423C16.6706 10.3685 16.6573 10.3936 16.6415 10.4171C16.6541 10.3923 16.6641 10.3664 16.6714 10.3397C16.6795 10.3126 16.69 10.2863 16.7027 10.261C16.7704 10.2048 16.8349 10.1453 16.896 10.0825C16.924 10.0522 16.9457 10.017 16.9599 9.97885C16.9659 9.95994 16.9704 9.94065 16.9735 9.92111C16.9762 9.90151 16.9762 9.88166 16.9735 9.86206C16.9591 9.7797 16.9187 9.70363 16.8579 9.64424C16.7993 9.57732 16.73 9.51303 16.666 9.44742C16.6466 9.39965 16.6355 9.3491 16.6334 9.29783C16.6323 9.24728 16.6411 9.197 16.6592 9.14956C16.7084 9.07124 16.7498 8.98865 16.783 8.90287C16.7978 8.86068 16.8033 8.81601 16.7993 8.77165C16.7943 8.72835 16.7804 8.68643 16.7585 8.64831C16.7146 8.57471 16.6636 8.50529 16.6061 8.44099C16.5779 8.40505 16.5529 8.36686 16.5313 8.32683C16.509 8.28622 16.496 8.24147 16.4932 8.19561C16.4905 8.15062 16.4989 8.10565 16.5177 8.0644C16.5365 8.02427 16.5628 7.9878 16.5953 7.9568C16.6608 7.89822 16.7298 7.84344 16.8021 7.79278C16.8167 7.76924 16.8294 7.74468 16.8402 7.7193C16.8437 7.68637 16.8505 7.65385 16.8606 7.6222Z" fill="#FF2316" />
-            <path d="M17.3014 7.60514C17.281 7.62829 17.2623 7.65285 17.2456 7.67862C17.2287 7.70457 17.2096 7.72915 17.1885 7.7521C17.1014 7.79168 17.0165 7.83547 16.9341 7.88332C16.8953 7.90659 16.8612 7.93637 16.8334 7.97123C16.8079 8.00484 16.7903 8.04332 16.7817 8.08408C16.7731 8.17028 16.7911 8.25699 16.8334 8.33339L16.8919 8.46461C16.9138 8.50606 16.9299 8.5502 16.9395 8.59582C16.9551 8.64874 16.9579 8.70441 16.9477 8.75853C16.9367 8.81195 16.915 8.86277 16.8837 8.90812C16.82 8.99282 16.7475 9.07107 16.6674 9.14168C16.6372 9.18103 16.615 9.22556 16.6021 9.2729C16.5898 9.31914 16.5856 9.36705 16.5898 9.41461V9.40281C16.632 9.48547 16.6851 9.56814 16.7259 9.65605C16.7465 9.69804 16.7625 9.74203 16.7735 9.78727C16.7844 9.83481 16.7844 9.88407 16.7735 9.93161C16.7676 9.95547 16.7594 9.97876 16.749 10.0012C16.7354 10.0221 16.7232 10.0431 16.7082 10.0628C16.6755 10.1008 16.6364 10.1332 16.5925 10.1586C16.5044 10.2049 16.4136 10.2461 16.3204 10.282C16.2978 10.3031 16.2768 10.3259 16.2578 10.3502C16.2383 10.3744 16.2169 10.3972 16.1939 10.4184C16.214 10.3955 16.2326 10.3714 16.2497 10.3463C16.2662 10.3197 16.2854 10.2946 16.3068 10.2715C16.3966 10.2301 16.4839 10.1837 16.568 10.1324C16.6069 10.1077 16.641 10.0766 16.6687 10.0405C16.6807 10.0223 16.6916 10.0035 16.7014 9.98409C16.7093 9.96459 16.7153 9.94437 16.7191 9.92373C16.7288 9.83633 16.7098 9.74818 16.6646 9.6718C16.6238 9.58913 16.5694 9.50909 16.5286 9.4238C16.5217 9.36974 16.5259 9.31492 16.5408 9.2624C16.5553 9.20956 16.5797 9.15974 16.6129 9.11544C16.6892 9.04686 16.7589 8.97176 16.8211 8.89106C16.8492 8.85154 16.8687 8.80688 16.8783 8.75984C16.8875 8.71656 16.8875 8.67192 16.8783 8.62863C16.8548 8.54218 16.8225 8.45822 16.7817 8.378C16.7631 8.33561 16.7485 8.29169 16.7381 8.24679C16.7279 8.20006 16.7279 8.1518 16.7381 8.10507C16.7489 8.057 16.7713 8.01205 16.8034 7.97386C16.8368 7.93766 16.8758 7.90667 16.9191 7.88201C17.0062 7.83346 17.1 7.80065 17.1912 7.76129C17.2139 7.74013 17.2348 7.71732 17.2538 7.69305C17.2671 7.66254 17.2831 7.63313 17.3014 7.60514Z" fill="#FF2316" />
-            <path d="M15.8374 10.1376C15.8479 10.0964 15.8706 10.059 15.9027 10.03C15.9376 10.0009 15.9765 9.97659 16.0183 9.95785C16.1021 9.92802 16.1813 9.88742 16.2537 9.83713C16.2889 9.80052 16.3125 9.755 16.3218 9.70591C16.3313 9.65485 16.3341 9.60282 16.3299 9.55108C16.3299 9.49728 16.3299 9.44217 16.3299 9.38706C16.331 9.3299 16.3444 9.27359 16.3694 9.22173C16.3855 9.17115 16.4156 9.12577 16.4565 9.09051C16.4989 9.05484 16.5491 9.02883 16.6034 9.0144C16.7078 8.98878 16.8137 8.96906 16.9204 8.95536C17.0019 8.92777 17.0692 8.87089 17.1082 8.79658C17.1194 8.71877 17.1217 8.64 17.115 8.56171C17.1078 8.47965 17.1128 8.39701 17.13 8.31633C17.1399 8.2747 17.1584 8.23543 17.1844 8.20086C17.1906 8.19171 17.1979 8.18334 17.2062 8.17593C17.2132 8.16744 17.2209 8.15953 17.2293 8.15231C17.2478 8.13793 17.2679 8.12561 17.2892 8.11557C17.3735 8.07752 17.4661 8.0644 17.545 8.02766C17.5986 7.99045 17.6417 7.94081 17.6702 7.88332C17.6577 7.91289 17.6426 7.94142 17.6253 7.96861C17.6073 7.99687 17.5837 8.02141 17.5559 8.04078C17.4769 8.08408 17.3844 8.10507 17.3109 8.14444C17.2929 8.15353 17.2761 8.16453 17.2606 8.17724C17.2606 8.17724 17.2484 8.19168 17.2416 8.19955C17.2349 8.20641 17.229 8.21388 17.2239 8.22186C17.2031 8.25399 17.1888 8.2896 17.1817 8.32683C17.1701 8.40472 17.1701 8.48382 17.1817 8.56171C17.1905 8.64372 17.1905 8.72639 17.1817 8.80839C17.1594 8.85224 17.1284 8.89146 17.0905 8.92386C17.0515 8.95888 17.0051 8.9853 16.9545 9.00128C16.8505 9.01468 16.7474 9.03352 16.6456 9.05771C16.6 9.06881 16.5576 9.08991 16.5218 9.11938C16.4884 9.14752 16.4635 9.18377 16.4497 9.22435C16.4259 9.26964 16.4125 9.3193 16.4102 9.37C16.4102 9.42118 16.4102 9.47497 16.4102 9.52877C16.4055 9.59239 16.3918 9.65509 16.3694 9.7151C16.3571 9.77216 16.3272 9.82431 16.2837 9.86469C16.2464 9.89193 16.2051 9.9136 16.1612 9.92898C16.1177 9.94604 16.0728 9.96047 16.0251 9.97622C15.9831 9.99096 15.9437 10.0117 15.9081 10.0379C15.8757 10.0644 15.8513 10.0989 15.8374 10.1376Z" fill="#FFD81F" />
-            <path d="M15.2836 10.0681C15.3109 10.0229 15.3504 9.98575 15.3979 9.96048C15.4483 9.93545 15.5025 9.91819 15.5585 9.9093C15.6679 9.89627 15.7753 9.87027 15.8782 9.83188C15.933 9.79832 15.9768 9.75057 16.0047 9.6941C16.0331 9.63765 16.0546 9.5782 16.0687 9.51696C16.085 9.45529 16.1 9.38575 16.1231 9.32801C16.1454 9.262 16.181 9.2009 16.2279 9.14824C16.2636 9.09248 16.3144 9.04713 16.3748 9.01703C16.4376 8.98851 16.5055 8.97202 16.5748 8.96848C16.7066 8.96633 16.8383 8.97158 16.9694 8.98422C17.0788 8.97692 17.1814 8.93006 17.2565 8.85301L17.2497 8.86088C17.2895 8.77683 17.3196 8.68883 17.3395 8.59845C17.36 8.50161 17.3944 8.408 17.4416 8.32027C17.4688 8.27512 17.5043 8.2351 17.5463 8.20217C17.5568 8.19315 17.5682 8.18523 17.5804 8.17855C17.5921 8.17079 17.6044 8.16378 17.6171 8.15756C17.6436 8.14568 17.6714 8.13687 17.7001 8.13132C17.813 8.10901 17.9246 8.1182 18.0307 8.09589C18.1063 8.0677 18.1738 8.02275 18.228 7.96467C18.2026 7.99509 18.1754 8.02401 18.1464 8.05127C18.1146 8.07839 18.0782 8.10013 18.0389 8.11557C17.9314 8.14575 17.8157 8.14444 17.7151 8.17068C17.6904 8.176 17.6667 8.18438 17.6443 8.19561C17.6338 8.20141 17.6238 8.20799 17.6144 8.2153C17.6048 8.22317 17.594 8.22973 17.5858 8.2376C17.5497 8.2686 17.5201 8.30602 17.4987 8.34782C17.4574 8.43106 17.4281 8.51936 17.4116 8.61026C17.3942 8.70754 17.3655 8.80262 17.3259 8.89368C17.2839 8.93887 17.2336 8.97623 17.1776 9.00391C17.1181 9.03397 17.0528 9.05184 16.9858 9.05639C16.8582 9.04482 16.7302 9.03956 16.6021 9.04065C16.5438 9.04255 16.4864 9.05548 16.4333 9.0787C16.385 9.10289 16.3435 9.13801 16.3122 9.18105C16.2695 9.22684 16.2367 9.28036 16.2156 9.33851C16.1925 9.39624 16.1762 9.45792 16.1571 9.5209C16.1314 9.58912 16.0976 9.65428 16.0564 9.7151C16.021 9.7773 15.9682 9.82867 15.9041 9.86337C15.8506 9.88499 15.7943 9.89913 15.7367 9.90536L15.5666 9.92636C15.512 9.93247 15.4584 9.94572 15.4074 9.96572C15.3569 9.98783 15.314 10.0233 15.2836 10.0681Z" fill="#FFD81F" />
-            <path d="M14.7353 9.78989C14.7813 9.74735 14.8381 9.71706 14.8999 9.70198C14.9655 9.68801 15.0329 9.68315 15.0999 9.68754C15.2283 9.70203 15.358 9.70203 15.4863 9.68754C15.5592 9.66442 15.6248 9.62377 15.6768 9.56945C15.729 9.51511 15.7747 9.45524 15.8129 9.391C15.8523 9.32801 15.8904 9.25978 15.9381 9.19811C15.9865 9.13093 16.0479 9.07344 16.119 9.02884C16.1791 8.9787 16.2511 8.94356 16.3286 8.92649C16.4086 8.91205 16.4907 8.91205 16.5708 8.92649C16.7195 8.9583 16.8662 8.99816 17.0102 9.0459C17.1388 9.06481 17.27 9.03875 17.3803 8.97241H17.3708C17.441 8.89042 17.5021 8.80158 17.5531 8.70736C17.6101 8.60589 17.6811 8.51232 17.764 8.42918C17.8103 8.3858 17.8642 8.3507 17.9232 8.32552C17.9382 8.31883 17.9537 8.31313 17.9695 8.30846C17.9855 8.30314 18.0019 8.29876 18.0185 8.29534C18.0524 8.29143 18.0866 8.29143 18.1205 8.29534C18.2566 8.29534 18.379 8.33864 18.5056 8.34258C18.6024 8.33075 18.6953 8.29849 18.7777 8.2481C18.7384 8.27444 18.6975 8.29854 18.6553 8.32027C18.6104 8.34211 18.5621 8.35674 18.5124 8.36357C18.3763 8.36357 18.2511 8.33733 18.1273 8.33995C18.098 8.33961 18.0688 8.3427 18.0402 8.34914C18.0262 8.35257 18.0125 8.35741 17.9994 8.36357C17.985 8.36734 17.9713 8.37308 17.9586 8.38063C17.9081 8.40561 17.8621 8.43841 17.8225 8.47773C17.7475 8.56043 17.684 8.65218 17.6334 8.75066C17.5789 8.85258 17.5137 8.94887 17.4388 9.03802C17.3757 9.07623 17.3062 9.10375 17.2334 9.11938C17.1567 9.13715 17.0771 9.13983 16.9994 9.12725C16.8511 9.08788 16.7109 9.03409 16.5708 9.01047C16.5038 8.998 16.435 8.998 16.368 9.01047C16.3055 9.02541 16.247 9.05316 16.1966 9.09182C16.1331 9.13052 16.0777 9.18037 16.0333 9.23878C15.9884 9.29652 15.9476 9.3595 15.8973 9.42249C15.8555 9.48913 15.8057 9.55079 15.7489 9.60619C15.6872 9.66559 15.6102 9.70815 15.5258 9.72953C15.4581 9.73805 15.3894 9.73805 15.3217 9.72953C15.2537 9.72953 15.1856 9.71379 15.1231 9.70854C15.059 9.69934 14.9939 9.69934 14.9298 9.70854C14.8583 9.71721 14.7908 9.74544 14.7353 9.78989Z" fill="#FFD81F" />
-            <path d="M14.6047 9.34507C14.6604 9.31482 14.723 9.29815 14.787 9.29652C14.8532 9.29559 14.9192 9.30399 14.9829 9.32145C15.1044 9.35962 15.2299 9.38471 15.3571 9.39624C15.4334 9.38896 15.5068 9.36373 15.5707 9.32276C15.6341 9.28145 15.6929 9.23396 15.7462 9.18105C15.8006 9.12987 15.8551 9.07476 15.9177 9.0249C15.9818 8.97124 16.0556 8.92943 16.1354 8.90156C16.2062 8.86674 16.2845 8.84832 16.3639 8.84776C16.4439 8.84925 16.5228 8.86527 16.5966 8.895C16.7318 8.95308 16.8632 9.01925 16.9898 9.09313C17.1111 9.1361 17.2442 9.1361 17.3654 9.09313H17.3545C17.4477 9.0292 17.5339 8.95621 17.6116 8.87531C17.6926 8.79126 17.784 8.71722 17.8838 8.65487C17.9396 8.62394 18.0003 8.6018 18.0634 8.58926C18.0797 8.58799 18.0961 8.58799 18.1124 8.58926C18.1291 8.58796 18.1459 8.58796 18.1627 8.58926C18.1967 8.58957 18.2305 8.59398 18.2634 8.60238C18.3913 8.63388 18.5001 8.69555 18.6212 8.7231C18.7177 8.73049 18.8146 8.71797 18.9056 8.68636C18.8615 8.70455 18.816 8.71946 18.7695 8.73098C18.7207 8.74331 18.6702 8.74774 18.6199 8.7441C18.4933 8.72179 18.375 8.66799 18.2566 8.64568C18.2275 8.63946 18.1979 8.63638 18.1681 8.6365C18.1545 8.6365 18.1396 8.6365 18.126 8.6365C18.1118 8.63837 18.0977 8.641 18.0838 8.64437C18.0277 8.65745 17.9743 8.67964 17.926 8.70998C17.8281 8.7707 17.7393 8.84398 17.662 8.9278C17.5826 9.01382 17.4941 9.09165 17.398 9.16005C17.3277 9.18424 17.2536 9.19666 17.179 9.19679C17.1012 9.19864 17.0239 9.18525 16.9517 9.15743C16.8157 9.09182 16.6987 9.0144 16.5694 8.96454C16.5081 8.93846 16.4419 8.92462 16.3748 8.92386C16.3103 8.92452 16.2466 8.93886 16.1884 8.96585C16.1176 8.99004 16.0518 9.02601 15.9939 9.07214C15.934 9.11806 15.8782 9.17055 15.8197 9.22041C15.7616 9.27503 15.6973 9.32302 15.6278 9.36344C15.5535 9.4075 15.4687 9.43236 15.3816 9.43561C15.3145 9.43149 15.2482 9.41914 15.1843 9.39887C15.1217 9.37918 15.0618 9.35688 15.0006 9.33851C14.9399 9.31891 14.8767 9.30701 14.8128 9.30308C14.7408 9.29909 14.669 9.31358 14.6047 9.34507Z" fill="#FFD81F" />
-            <path d="M14.6237 8.9383C14.6832 8.919 14.7468 8.91405 14.8087 8.92386C14.8714 8.936 14.9323 8.95584 14.9897 8.98291C15.097 9.0392 15.2107 9.08323 15.3285 9.11413C15.4039 9.12137 15.4799 9.11102 15.5503 9.08395C15.6206 9.05705 15.688 9.02365 15.7517 8.98422C15.8156 8.94486 15.8877 8.90418 15.953 8.86875C16.0277 8.83103 16.1088 8.80658 16.1925 8.79658C16.2678 8.77638 16.347 8.77323 16.4238 8.7874C16.4993 8.80214 16.5709 8.83113 16.6347 8.87269C16.7501 8.9528 16.8597 9.04048 16.9626 9.13512C17.0697 9.1969 17.1954 9.22188 17.3191 9.20598H17.3082C17.4122 9.16204 17.5119 9.10935 17.6062 9.04852C17.705 8.98434 17.8114 8.93191 17.9232 8.89237C17.9834 8.87345 18.0463 8.86371 18.1096 8.8635H18.1586C18.1749 8.86223 18.1913 8.86223 18.2076 8.8635C18.2396 8.86996 18.2706 8.88009 18.3001 8.89368C18.4158 8.94617 18.5042 9.0249 18.6131 9.07083C18.7042 9.09556 18.7997 9.10138 18.8934 9.08788C18.8485 9.09663 18.803 9.10233 18.7573 9.10494C18.7083 9.10731 18.6592 9.10244 18.6117 9.09051C18.4961 9.04589 18.3967 8.97504 18.2892 8.93174C18.263 8.92145 18.2356 8.91397 18.2076 8.90943C18.1936 8.90814 18.1795 8.90814 18.1654 8.90943C18.1505 8.90943 18.1368 8.90943 18.1219 8.90943C18.0655 8.91172 18.0099 8.9228 17.9572 8.94223C17.8524 8.9827 17.7533 9.0356 17.662 9.09969C17.5656 9.16554 17.4631 9.22264 17.3558 9.27028H17.3436C17.2715 9.28139 17.198 9.28139 17.1259 9.27028C17.0524 9.25768 16.9821 9.23138 16.9191 9.19286C16.8089 9.10757 16.7123 9.01309 16.6007 8.94355C16.5475 8.90868 16.4879 8.88375 16.4252 8.87006C16.3633 8.85761 16.2993 8.85761 16.2374 8.87006C16.1637 8.87906 16.0922 8.90084 16.0265 8.93436C15.9585 8.96585 15.8904 9.00391 15.8251 9.04065C15.7563 9.08014 15.6839 9.11352 15.6088 9.14037C15.527 9.16717 15.4398 9.17482 15.3544 9.16268C15.2917 9.14615 15.2314 9.12236 15.1748 9.09182C15.119 9.06295 15.0673 9.03146 15.0128 9.00259C14.9592 8.9736 14.9021 8.95112 14.8428 8.93567C14.771 8.91652 14.695 8.91743 14.6237 8.9383Z" fill="#FFD81F" />
-            <path d="M14.5842 8.40294C14.6519 8.39832 14.7197 8.41004 14.7815 8.43705C14.8413 8.46444 14.8971 8.49929 14.9475 8.54071C15.039 8.62604 15.1403 8.70089 15.2496 8.76378C15.326 8.78948 15.4073 8.79845 15.4877 8.79002C15.5694 8.78046 15.65 8.76377 15.7285 8.74016C15.8074 8.71785 15.8891 8.69292 15.9748 8.67587C16.0639 8.65551 16.1559 8.64974 16.2469 8.65881C16.33 8.65742 16.4123 8.67401 16.4878 8.70736C16.5596 8.74135 16.6239 8.78861 16.6769 8.84645C16.7687 8.95334 16.8523 9.06648 16.9272 9.18499C17.0193 9.27543 17.1409 9.33245 17.2715 9.34638H17.2592C17.3812 9.32896 17.5014 9.3013 17.6184 9.26372C17.7415 9.22441 17.8689 9.19887 17.9981 9.18761C18.0662 9.18389 18.1346 9.19008 18.2008 9.20598L18.2498 9.22041C18.2661 9.22586 18.282 9.23243 18.2974 9.2401C18.328 9.25415 18.3564 9.27226 18.3818 9.2939C18.4838 9.37525 18.5491 9.47366 18.6457 9.55633C18.7331 9.6033 18.8293 9.63318 18.9287 9.64424C18.8784 9.64424 18.8267 9.63506 18.7777 9.62587C18.7267 9.61567 18.6776 9.59797 18.6321 9.57339C18.5273 9.4999 18.4498 9.4028 18.3518 9.33326C18.3286 9.3156 18.3029 9.30104 18.2756 9.28996C18.2622 9.28359 18.248 9.27874 18.2335 9.27553C18.2193 9.27065 18.2047 9.26713 18.1899 9.26503C18.1305 9.25223 18.0694 9.24913 18.009 9.25584C17.8872 9.27043 17.7675 9.29861 17.6525 9.33982C17.5308 9.38119 17.4056 9.41193 17.2783 9.43167H17.266C17.1889 9.42268 17.1136 9.40233 17.0429 9.37131C16.9715 9.34067 16.9076 9.29598 16.8551 9.2401C16.7771 9.12055 16.6922 9.00534 16.6007 8.895C16.5572 8.846 16.504 8.80585 16.4442 8.7769C16.3828 8.7496 16.316 8.73529 16.2483 8.73491C16.1684 8.726 16.0875 8.72955 16.0088 8.74541C15.9285 8.75984 15.8483 8.78215 15.7666 8.80183C15.6827 8.82487 15.5965 8.83982 15.5095 8.84645C15.4167 8.85376 15.3235 8.83982 15.2373 8.80577C15.1785 8.77473 15.1244 8.73586 15.0768 8.6903C15.0292 8.647 14.9856 8.60238 14.9407 8.55908C14.8954 8.5158 14.8441 8.47871 14.7883 8.44886C14.7253 8.4175 14.6552 8.40172 14.5842 8.40294Z" fill="#FFD81F" />
-            <path d="M15.0373 8.15231C15.0951 8.15977 15.1501 8.18047 15.1979 8.21267C15.2429 8.24494 15.2828 8.2834 15.3163 8.32683C15.3758 8.41311 15.4467 8.49162 15.5272 8.56039C15.5872 8.59488 15.6547 8.61554 15.7245 8.62076C15.7964 8.62664 15.8688 8.62664 15.9408 8.62076C16.0129 8.62076 16.0877 8.60763 16.1639 8.60763C16.2432 8.6048 16.3223 8.61501 16.398 8.63781C16.4682 8.64934 16.5348 8.67628 16.5925 8.71654C16.6479 8.75596 16.6942 8.806 16.7286 8.86351C16.7813 8.96811 16.8267 9.07594 16.8647 9.1863C16.9241 9.2778 17.016 9.3454 17.1232 9.37656H17.1123C17.2203 9.38261 17.3287 9.37954 17.4361 9.36738C17.5487 9.35554 17.6623 9.35554 17.7749 9.36738C17.8334 9.3751 17.8903 9.39147 17.9436 9.41593C17.9568 9.4217 17.9695 9.42827 17.9817 9.43561C17.9945 9.44272 18.0068 9.45061 18.0185 9.45923C18.0412 9.47654 18.0617 9.49636 18.0797 9.51828C18.1477 9.60225 18.1831 9.69542 18.2484 9.78071C18.3122 9.83488 18.3874 9.87516 18.4688 9.8988C18.4269 9.88786 18.3855 9.87472 18.345 9.85944C18.3042 9.84252 18.2662 9.81998 18.2321 9.79252C18.1573 9.71379 18.1124 9.62062 18.0443 9.54714C18.0281 9.52861 18.0093 9.51228 17.9885 9.49859C17.9786 9.49103 17.9676 9.48485 17.9559 9.48022C17.9453 9.47326 17.9339 9.46753 17.9219 9.46316C17.8732 9.44373 17.8219 9.43135 17.7695 9.42642C17.6619 9.41926 17.5538 9.42322 17.447 9.43823C17.3352 9.45324 17.2224 9.45895 17.1096 9.45529C17.0463 9.43596 16.9867 9.40718 16.9327 9.37C16.8777 9.33339 16.8313 9.28599 16.7966 9.23091C16.757 9.12373 16.7116 9.01862 16.6606 8.91599C16.6331 8.86836 16.5961 8.82644 16.5517 8.79265C16.5048 8.75982 16.4509 8.73742 16.3939 8.72704C16.3226 8.69217 16.245 8.67121 16.1653 8.66537C16.0932 8.66537 16.0197 8.66537 15.9462 8.67324C15.8702 8.67976 15.7937 8.67976 15.7176 8.67324C15.6389 8.66439 15.5634 8.63788 15.4972 8.59582C15.4535 8.56024 15.4155 8.51868 15.3843 8.47248C15.353 8.42918 15.3258 8.38456 15.2959 8.34126C15.2661 8.29787 15.2304 8.25859 15.1897 8.22448C15.1452 8.18964 15.0931 8.16496 15.0373 8.15231Z" fill="#FFD81F" />
-            <path d="M15.3339 7.82427C15.3862 7.84353 15.4328 7.87458 15.47 7.91481C15.5042 7.95438 15.5322 7.9986 15.553 8.04603C15.5861 8.13895 15.6318 8.22721 15.6891 8.30846C15.7379 8.35516 15.7976 8.38982 15.8632 8.4095C15.9319 8.43066 16.0025 8.44559 16.0741 8.45411C16.1462 8.46461 16.2197 8.47379 16.2946 8.48954C16.3714 8.50523 16.4453 8.53223 16.5136 8.56958C16.5762 8.59437 16.6318 8.63328 16.6756 8.68306C16.7194 8.73283 16.7502 8.79201 16.7653 8.85563C16.786 8.96623 16.7991 9.078 16.8048 9.19023C16.8372 9.29053 16.9077 9.37534 17.0021 9.42774H16.9912C17.0954 9.45671 17.2019 9.47734 17.3096 9.48941C17.4222 9.50153 17.5331 9.52574 17.6402 9.56158C17.6945 9.58103 17.745 9.60896 17.7899 9.64424C17.8016 9.65202 17.8121 9.66129 17.8212 9.6718C17.8322 9.68086 17.8422 9.69097 17.8511 9.70198C17.8679 9.72303 17.8821 9.74593 17.8933 9.77021C17.9355 9.866 17.9423 9.96179 17.9831 10.051C18.0307 10.1167 18.0924 10.1717 18.1641 10.2124C18.1265 10.1926 18.0902 10.1707 18.0552 10.1468C18.02 10.123 17.9896 10.0932 17.9654 10.0589C17.9164 9.96835 17.8987 9.86994 17.8552 9.78596C17.8445 9.76487 17.8312 9.74505 17.8157 9.72691C17.8089 9.71772 17.7967 9.71116 17.7885 9.70198C17.7796 9.69384 17.7701 9.68638 17.76 9.67967C17.7185 9.6504 17.6726 9.62739 17.6239 9.61144C17.5207 9.58079 17.4144 9.56099 17.3069 9.55239C17.1943 9.54209 17.0828 9.52277 16.9735 9.49466C16.9183 9.46209 16.8687 9.42135 16.8266 9.37394C16.7851 9.32781 16.7552 9.27301 16.7395 9.21385C16.7332 9.1042 16.7196 8.99506 16.6987 8.88712C16.6868 8.83547 16.6631 8.78704 16.6293 8.74541C16.592 8.70368 16.5455 8.67051 16.4932 8.64831C16.4336 8.61434 16.3689 8.58955 16.3014 8.57483C16.232 8.55908 16.1653 8.54859 16.0864 8.53546C16.0114 8.52344 15.9376 8.50499 15.866 8.48035C15.7926 8.45549 15.7267 8.41358 15.6741 8.35832C15.6422 8.31551 15.6174 8.26813 15.6006 8.21792C15.5829 8.16937 15.5693 8.12213 15.5517 8.07489C15.5352 8.02816 15.5118 7.98396 15.4823 7.94368C15.446 7.89073 15.3944 7.84922 15.3339 7.82427Z" fill="#FFD81F" />
-            <path d="M15.8156 7.79015C15.8564 7.8146 15.8897 7.849 15.9122 7.88988C15.9323 7.92918 15.9461 7.97122 15.953 8.01453C15.9596 8.0979 15.9779 8.18001 16.0075 8.2586C16.0387 8.30575 16.0818 8.34456 16.1326 8.37144C16.1866 8.4002 16.243 8.42435 16.3014 8.44361C16.3599 8.46461 16.4211 8.48429 16.481 8.51053C16.5422 8.53549 16.5988 8.57008 16.6483 8.61288C16.6975 8.64678 16.7372 8.69185 16.764 8.7441C16.7877 8.79198 16.7993 8.84458 16.798 8.89762C16.7928 8.99384 16.7796 9.0895 16.7585 9.18367C16.7673 9.26707 16.8081 9.34429 16.8728 9.40018C16.9544 9.442 17.0395 9.47712 17.1273 9.50515C17.2201 9.53356 17.3091 9.57231 17.3926 9.62062C17.4336 9.64561 17.47 9.67711 17.5001 9.71379C17.5081 9.72286 17.5154 9.73251 17.5218 9.74266C17.5285 9.7518 17.5344 9.76146 17.5395 9.77152C17.5486 9.79222 17.5555 9.81376 17.5599 9.83582C17.5735 9.92111 17.5599 10.0025 17.5735 10.0825C17.5985 10.145 17.6387 10.2008 17.6906 10.2452C17.6626 10.2229 17.6363 10.1988 17.6116 10.173C17.588 10.1475 17.5692 10.1182 17.5559 10.0864C17.5355 10.0038 17.5423 9.9198 17.5232 9.84107C17.5191 9.82188 17.5122 9.80333 17.5028 9.78596C17.498 9.77725 17.4926 9.76892 17.4865 9.76103C17.4865 9.75315 17.4742 9.74528 17.4661 9.73741C17.4372 9.70601 17.4032 9.67941 17.3654 9.65868C17.2845 9.61508 17.1987 9.58073 17.1096 9.55633C17.0158 9.52934 16.9247 9.49421 16.8374 9.45135C16.7977 9.41628 16.765 9.37447 16.7408 9.32801C16.7157 9.28228 16.7022 9.23147 16.7014 9.17974C16.7207 9.08791 16.7338 8.995 16.7408 8.90156C16.7428 8.85635 16.7335 8.81134 16.7136 8.77034C16.6913 8.72947 16.6597 8.69402 16.6211 8.66668C16.5779 8.62794 16.5282 8.59645 16.4742 8.57352C16.4197 8.54859 16.3599 8.52759 16.3014 8.50397C16.2404 8.48149 16.1817 8.45382 16.1258 8.42131C16.0694 8.3898 16.0224 8.34456 15.9898 8.29009C15.9722 8.24824 15.9616 8.20394 15.9585 8.15887C15.9585 8.11688 15.9585 8.07489 15.9503 8.02766C15.9472 7.9857 15.9371 7.9445 15.9204 7.90562C15.8982 7.85793 15.8617 7.81772 15.8156 7.79015Z" fill="#FFD81F" />
-            <path d="M16.1054 7.48442C16.1425 7.52062 16.1684 7.56605 16.1803 7.61564C16.1891 7.66293 16.1891 7.71137 16.1803 7.75866C16.1619 7.84894 16.1569 7.94127 16.1653 8.0329C16.1857 8.09027 16.2208 8.14173 16.2673 8.18249C16.3179 8.22566 16.372 8.26472 16.4293 8.29927C16.4878 8.3347 16.549 8.37144 16.6061 8.41212C16.6652 8.45351 16.7167 8.50406 16.7585 8.56171C16.8016 8.60728 16.8316 8.66291 16.8456 8.7231C16.8569 8.7803 16.8536 8.83928 16.8361 8.895C16.7982 8.99689 16.7527 9.096 16.7 9.19155C16.6893 9.238 16.6885 9.28609 16.6979 9.33282C16.7073 9.37955 16.7265 9.42393 16.7544 9.46316L16.7476 9.4566C16.825 9.52013 16.9079 9.57719 16.9953 9.62719C17.087 9.67955 17.1723 9.7416 17.2497 9.8122C17.2874 9.85055 17.3178 9.89496 17.3395 9.94342C17.3395 9.95523 17.349 9.96704 17.3531 9.97885C17.3582 9.99066 17.3618 10.003 17.364 10.0156C17.3675 10.0395 17.3675 10.0638 17.364 10.0878C17.3531 10.1835 17.3109 10.2662 17.3028 10.3502C17.3103 10.4232 17.337 10.4932 17.3803 10.5536C17.3573 10.5228 17.3359 10.4908 17.3164 10.4578C17.2985 10.4246 17.287 10.3886 17.2824 10.3515C17.2824 10.2583 17.3191 10.1691 17.3218 10.0812C17.3252 10.0603 17.3252 10.0391 17.3218 10.0182C17.3194 10.0074 17.3158 9.99679 17.3109 9.98672C17.3109 9.97753 17.3028 9.96704 17.2973 9.95654C17.2748 9.91618 17.2454 9.8798 17.2103 9.84894C17.1347 9.78357 17.0516 9.72677 16.9626 9.67967C16.8697 9.6288 16.7814 9.57041 16.6987 9.50515C16.6663 9.45802 16.6432 9.40553 16.6306 9.35032C16.6163 9.29562 16.6163 9.23837 16.6306 9.18367C16.6816 9.08998 16.727 8.99361 16.7667 8.895C16.7813 8.84867 16.7845 8.79972 16.7762 8.75197C16.7644 8.70349 16.7406 8.65848 16.7068 8.62076C16.6694 8.56996 16.6234 8.5256 16.5708 8.48954C16.5191 8.45017 16.4606 8.41343 16.402 8.37538C16.343 8.33614 16.2875 8.29226 16.236 8.24416C16.1855 8.1959 16.1489 8.13574 16.1299 8.06965C16.1217 8.02142 16.1217 7.97222 16.1299 7.92399C16.1381 7.87676 16.1503 7.83214 16.1585 7.79278C16.166 7.74714 16.166 7.70064 16.1585 7.655C16.1602 7.59413 16.1416 7.53435 16.1054 7.48442Z" fill="#FFD81F" />
-            <path d="M16.5191 7.34402C16.5469 7.38729 16.5611 7.43744 16.5599 7.48836C16.557 7.53839 16.5455 7.58761 16.5259 7.63401C16.481 7.72012 16.4494 7.81206 16.432 7.90694C16.4365 7.96882 16.4586 8.0283 16.4959 8.07883C16.5354 8.13421 16.581 8.18526 16.632 8.23104C16.6823 8.27959 16.734 8.32814 16.7803 8.38194C16.8303 8.43501 16.8691 8.49696 16.8946 8.56433C16.9257 8.61981 16.9408 8.68237 16.9381 8.74541C16.9334 8.80603 16.9133 8.86461 16.8796 8.91599C16.8135 9.0107 16.7403 9.10061 16.6606 9.18499C16.6364 9.22809 16.622 9.27561 16.6183 9.32446C16.6145 9.3733 16.6215 9.42236 16.6388 9.46841V9.46054C16.7009 9.54092 16.7701 9.61599 16.8456 9.68492C16.925 9.75635 16.995 9.83701 17.0538 9.92505C17.0825 9.96905 17.102 10.0181 17.1109 10.0694C17.1129 10.082 17.1129 10.0948 17.1109 10.1074C17.113 10.12 17.113 10.1329 17.1109 10.1455C17.1081 10.1711 17.1017 10.1963 17.0919 10.2203C17.0538 10.3134 16.9871 10.3869 16.9558 10.4748C16.9439 10.5486 16.9519 10.624 16.979 10.694C16.963 10.6588 16.9503 10.6224 16.9409 10.5851C16.9334 10.5469 16.9334 10.5078 16.9409 10.4696C16.9694 10.3751 17.0293 10.2938 17.0579 10.2072C17.0653 10.186 17.0695 10.1639 17.0701 10.1416C17.0715 10.1307 17.0715 10.1196 17.0701 10.1087C17.0715 10.0979 17.0715 10.0868 17.0701 10.0759C17.0593 10.0291 17.0394 9.98457 17.0116 9.94473C16.9466 9.85936 16.8707 9.78221 16.7857 9.7151C16.7051 9.64482 16.6313 9.56751 16.5653 9.48416C16.5459 9.42931 16.5372 9.37149 16.5395 9.31358C16.5417 9.25469 16.5595 9.19734 16.5912 9.14693C16.6678 9.06335 16.7396 8.97574 16.8062 8.8845C16.8345 8.84185 16.8518 8.79324 16.8565 8.74279C16.8583 8.6918 16.8466 8.64121 16.8225 8.59582C16.8008 8.53611 16.7675 8.4809 16.7245 8.43312C16.6823 8.38194 16.632 8.33339 16.5885 8.28222C16.5375 8.23161 16.4919 8.17622 16.4524 8.11688C16.4132 8.05814 16.392 7.98991 16.3912 7.92006C16.3991 7.86985 16.4152 7.82115 16.4388 7.77572C16.4606 7.73111 16.4864 7.68781 16.5082 7.6445C16.5303 7.60337 16.5455 7.55907 16.5531 7.51329C16.5618 7.45478 16.5498 7.39514 16.5191 7.34402Z" fill="#FFD81F" />
-            <path d="M17.0592 7.09077C17.0769 7.1468 17.0769 7.20663 17.0592 7.26266C17.0379 7.3177 17.0076 7.36909 16.9694 7.41488C16.8896 7.49865 16.8223 7.59271 16.7694 7.69437C16.7536 7.76377 16.7583 7.83606 16.783 7.903C16.8093 7.97392 16.8431 8.04207 16.8837 8.10639C16.9245 8.17331 16.9667 8.2376 17.0021 8.31371C17.0403 8.38619 17.0638 8.465 17.0715 8.54596C17.0877 8.61656 17.0835 8.6901 17.0592 8.75853C17.0325 8.82549 16.9896 8.88533 16.9341 8.93305C16.8258 9.02204 16.7112 9.1036 16.5912 9.17711C16.5493 9.21775 16.5167 9.26644 16.4956 9.32003C16.4745 9.37361 16.4653 9.43087 16.4687 9.4881V9.47629C16.5113 9.58253 16.5647 9.68449 16.6279 9.78071C16.6939 9.88063 16.7465 9.98821 16.7844 10.1009C16.8015 10.1579 16.8066 10.2177 16.7993 10.2767C16.8008 10.2915 16.8008 10.3065 16.7993 10.3213C16.7961 10.3361 16.7915 10.3506 16.7857 10.3646C16.7744 10.3925 16.7588 10.4186 16.7395 10.442C16.6646 10.5352 16.5626 10.6008 16.4932 10.69C16.4544 10.768 16.4375 10.8544 16.4442 10.9407C16.4401 10.897 16.4401 10.8531 16.4442 10.8094C16.4465 10.7641 16.4576 10.7196 16.4769 10.6782C16.5408 10.5811 16.6388 10.505 16.7014 10.4158C16.7168 10.3943 16.7292 10.3709 16.7381 10.3463C16.7423 10.3343 16.745 10.322 16.7463 10.3095C16.7463 10.2964 16.7463 10.2833 16.7463 10.2701C16.7485 10.2173 16.7412 10.1644 16.7245 10.114C16.6831 10.0076 16.6288 9.90626 16.5626 9.8122C16.4944 9.71273 16.4361 9.60729 16.3884 9.49728V9.47891C16.3849 9.41215 16.3946 9.34534 16.417 9.28209C16.4402 9.21606 16.4799 9.15656 16.5327 9.10888C16.6487 9.03575 16.7605 8.95646 16.8674 8.87138C16.9142 8.8318 16.9509 8.78243 16.9749 8.72704C16.994 8.67121 16.9982 8.61165 16.9871 8.55383C16.9827 8.4823 16.9638 8.4123 16.9313 8.34782C16.9 8.27959 16.8592 8.21661 16.8225 8.14313C16.7821 8.07397 16.7493 8.00098 16.7245 7.92531C16.7007 7.84825 16.7007 7.76617 16.7245 7.68912C16.7508 7.63409 16.7861 7.58356 16.8293 7.53953C16.8701 7.49492 16.9136 7.45424 16.9531 7.40831C16.9934 7.36693 17.0265 7.31952 17.0511 7.26791C17.0721 7.21099 17.075 7.14927 17.0592 7.09077Z" fill="#FFD81F" />
-            <path d="M17.3803 7.52772C17.3815 7.57397 17.3688 7.61955 17.3436 7.65894C17.3158 7.70064 17.2808 7.73746 17.2402 7.76785C17.159 7.82244 17.0858 7.88731 17.0225 7.96074C16.995 8.01246 16.9845 8.07107 16.9926 8.12869C16.9988 8.18922 17.013 8.24874 17.0347 8.30583C17.0538 8.36488 17.0756 8.42524 17.0905 8.48954C17.1067 8.55307 17.1099 8.61902 17.1 8.68374C17.0999 8.74294 17.0819 8.80084 17.0483 8.85038C17.0127 8.90011 16.9663 8.94175 16.9123 8.97241C16.8062 9.027 16.6967 9.0752 16.5844 9.11675C16.5428 9.14244 16.5071 9.17606 16.4795 9.21555C16.4519 9.25504 16.4329 9.29957 16.4238 9.34638V9.33457C16.4388 9.424 16.4625 9.51186 16.4946 9.59701C16.529 9.68692 16.5505 9.78095 16.5585 9.8765C16.5614 9.9251 16.554 9.97377 16.5367 10.0195C16.5329 10.0309 16.5279 10.0419 16.5218 10.0523C16.5168 10.0639 16.5104 10.0749 16.5027 10.0851C16.4879 10.1059 16.47 10.1244 16.4497 10.1402C16.3694 10.2045 16.2741 10.2413 16.1993 10.3016C16.1517 10.3568 16.1206 10.4235 16.1095 10.4945C16.1129 10.4586 16.1197 10.423 16.1299 10.3882C16.141 10.3525 16.159 10.3191 16.183 10.2898C16.2551 10.2229 16.349 10.177 16.417 10.1153C16.4341 10.1005 16.4488 10.0832 16.4606 10.0641C16.4666 10.0551 16.4716 10.0454 16.4755 10.0353C16.4802 10.0256 16.4838 10.0155 16.4864 10.0051C16.4992 9.96251 16.5033 9.91796 16.4986 9.87387C16.487 9.78365 16.4628 9.69536 16.4265 9.61144C16.3904 9.52202 16.364 9.42924 16.3476 9.33457V9.32539C16.3573 9.27061 16.3786 9.21834 16.4102 9.17186C16.4431 9.123 16.4867 9.08177 16.5381 9.05115C16.6483 9.00259 16.7626 8.96585 16.8579 8.91205C16.9039 8.88732 16.9435 8.85276 16.9735 8.81102C16.9999 8.7718 17.0149 8.7265 17.0171 8.6798C17.0292 8.6231 17.0292 8.56461 17.0171 8.50791C17.0048 8.44886 16.9858 8.38981 16.9681 8.32814C16.9492 8.26657 16.9374 8.20322 16.9327 8.13919C16.9277 8.07371 16.9433 8.0083 16.9776 7.95155C17.0102 7.91177 17.0487 7.87684 17.0919 7.84789C17.1341 7.81771 17.1776 7.79147 17.2184 7.7626C17.26 7.73622 17.2968 7.70341 17.3273 7.6655C17.3582 7.62538 17.3766 7.57758 17.3803 7.52772Z" fill="#FFD81F" />
-            <path d="M15.8197 10.5496L15.8469 10.4932C15.8592 10.4762 15.87 10.4578 15.8836 10.4407C15.9114 10.4066 15.945 10.3773 15.983 10.3541C16.0622 10.3099 16.144 10.27 16.2279 10.2347C16.2659 10.2157 16.3012 10.1919 16.3327 10.1639C16.3463 10.1494 16.3612 10.135 16.3735 10.1192C16.3856 10.1034 16.3965 10.0868 16.4061 10.0694C16.4482 10.0023 16.4686 9.92463 16.4646 9.84632C16.4515 9.76554 16.4268 9.68693 16.3912 9.61275C16.3737 9.57053 16.3614 9.52649 16.3544 9.48153C16.3466 9.43812 16.3466 9.39373 16.3544 9.35032C16.3728 9.26249 16.4167 9.1815 16.481 9.11675C16.5504 9.05283 16.6335 9.00445 16.7245 8.97504C16.8056 8.94981 16.8799 8.9077 16.9422 8.8517C16.9972 8.79602 17.0357 8.72715 17.0538 8.65225C17.0695 8.57604 17.0614 8.49705 17.0307 8.42524C16.9945 8.34602 16.9711 8.26195 16.9613 8.17593C16.9593 8.0882 16.9844 8.0019 17.0334 7.92793C17.0578 7.89 17.088 7.85593 17.1232 7.82689C17.1588 7.79744 17.1987 7.77313 17.2416 7.75473C17.3273 7.71667 17.4184 7.69568 17.4946 7.65238C17.5704 7.61089 17.632 7.54883 17.6715 7.47392L17.6443 7.52903C17.6321 7.5474 17.6212 7.56577 17.6089 7.58152C17.5804 7.61577 17.5464 7.64544 17.5082 7.66944C17.432 7.7193 17.3409 7.74685 17.2647 7.78753C17.2269 7.80701 17.1918 7.8308 17.1599 7.85839C17.1452 7.8723 17.1316 7.88721 17.1191 7.903C17.0596 7.98175 17.0285 8.07718 17.0307 8.17462C17.0442 8.25545 17.0694 8.33406 17.1055 8.40818C17.1217 8.45069 17.1336 8.49464 17.1409 8.5394C17.1482 8.58286 17.1482 8.62716 17.1409 8.67062C17.121 8.75834 17.0774 8.8394 17.0143 8.90549C16.9437 8.96854 16.8603 9.01679 16.7694 9.04721C16.6886 9.07228 16.6143 9.11391 16.5517 9.16924C16.4966 9.22464 16.4584 9.29367 16.4415 9.36869C16.432 9.40511 16.4292 9.44287 16.4333 9.48022C16.4396 9.51952 16.4496 9.55817 16.4633 9.59569C16.4994 9.67537 16.5228 9.75987 16.5327 9.84632C16.5368 9.93415 16.513 10.0211 16.4646 10.0956C16.4537 10.114 16.4374 10.1311 16.4252 10.1481C16.4129 10.1652 16.3925 10.1809 16.3762 10.1967C16.3395 10.225 16.2993 10.2488 16.2565 10.2675C16.1721 10.3056 16.0809 10.3279 16.0034 10.3712C15.9251 10.4117 15.8611 10.4738 15.8197 10.5496Z" fill="#20C20E" />
-            <path d="M15.2088 10.3239L15.2578 10.2741C15.2768 10.2596 15.2945 10.2439 15.3149 10.2308C15.357 10.2041 15.4028 10.1833 15.451 10.1691C15.5489 10.1389 15.6551 10.1363 15.7503 10.1153C15.7973 10.1063 15.843 10.0918 15.8864 10.072C15.9065 10.0616 15.926 10.0502 15.9449 10.0379C15.9634 10.0254 15.9811 10.0118 15.9979 9.99722C16.0634 9.94035 16.1106 9.86658 16.134 9.78464C16.1504 9.69806 16.1531 9.60959 16.1422 9.52221C16.1395 9.47526 16.1422 9.42817 16.1503 9.38181C16.1606 9.33539 16.179 9.29101 16.2048 9.25059C16.2582 9.1656 16.3337 9.0955 16.4238 9.04721C16.5202 9.00135 16.6253 8.97501 16.7327 8.96979C16.8269 8.966 16.9194 8.94414 17.0048 8.90549C17.0824 8.86419 17.148 8.80461 17.1953 8.73229C17.2389 8.66052 17.2593 8.57781 17.2538 8.49479C17.2454 8.40297 17.2518 8.31045 17.2728 8.22054C17.3037 8.12856 17.3619 8.04734 17.4402 7.98698C17.4595 7.97109 17.4799 7.95661 17.5014 7.94368C17.5232 7.93056 17.5477 7.92006 17.5708 7.90825C17.6195 7.88952 17.6702 7.87589 17.7219 7.86757C17.8266 7.85314 17.93 7.85576 18.0253 7.83346C18.1202 7.81207 18.2075 7.76679 18.2784 7.70224L18.2335 7.74685C18.2144 7.76129 18.1967 7.77703 18.1763 7.79015C18.1345 7.81723 18.0886 7.83803 18.0402 7.85183C17.9409 7.88201 17.8361 7.88463 17.7395 7.90562C17.6926 7.91488 17.6469 7.92942 17.6035 7.94893L17.5436 7.98304C17.5259 7.99748 17.5069 8.00929 17.4905 8.02503C17.4253 8.08154 17.3781 8.15483 17.3545 8.23629C17.3402 8.32322 17.3402 8.41179 17.3545 8.49872C17.3567 8.54616 17.3535 8.59368 17.345 8.64044C17.3347 8.68683 17.3163 8.73121 17.2905 8.77165C17.2379 8.8576 17.1622 8.92831 17.0715 8.97635C16.9747 9.02115 16.8698 9.04744 16.7626 9.05377C16.6683 9.05706 16.5757 9.07894 16.4905 9.11806C16.4133 9.15895 16.3482 9.2181 16.3014 9.28996C16.2786 9.32468 16.2621 9.36282 16.2524 9.40281C16.2434 9.44598 16.2397 9.49002 16.2415 9.53402C16.2499 9.62627 16.2434 9.71921 16.2224 9.80958C16.1912 9.90102 16.1331 9.98172 16.0551 10.0418C16.0358 10.0577 16.0154 10.0722 15.9939 10.0851C15.9716 10.0985 15.9484 10.1103 15.9245 10.1206C15.876 10.1401 15.8253 10.1537 15.7734 10.1612C15.67 10.177 15.5666 10.173 15.47 10.1954C15.3724 10.2144 15.282 10.2588 15.2088 10.3239Z" fill="#20C20E" />
-            <path d="M15.2605 9.81351L15.3108 9.78071L15.3639 9.75578C15.4032 9.74123 15.4444 9.73196 15.4863 9.72822C15.5707 9.71904 15.6551 9.72822 15.7367 9.72822C15.7761 9.72794 15.8153 9.72354 15.8537 9.7151L15.9081 9.69673C15.9245 9.68754 15.9422 9.6823 15.9585 9.6718C16.0222 9.6377 16.0749 9.5873 16.1109 9.52615C16.1399 9.46235 16.1588 9.39465 16.1667 9.32539C16.1747 9.28848 16.1866 9.25243 16.202 9.21779C16.2192 9.18143 16.2427 9.14816 16.2714 9.11938C16.3306 9.05905 16.4045 9.01395 16.4864 8.98816C16.5717 8.96709 16.6604 8.96219 16.7476 8.97373C16.8252 8.9843 16.9041 8.98074 16.9803 8.96323C17.05 8.94205 17.1134 8.90516 17.1654 8.85563C17.2139 8.80666 17.2465 8.74509 17.2592 8.67849C17.27 8.60458 17.2929 8.53281 17.3273 8.46592C17.3712 8.39866 17.4333 8.34422 17.5069 8.30846C17.5246 8.29796 17.545 8.29271 17.564 8.28353L17.6266 8.26647C17.6713 8.25758 17.7171 8.25449 17.7627 8.25728C17.8484 8.25728 17.93 8.27828 18.0117 8.27434C18.0922 8.27186 18.1706 8.24877 18.2389 8.20742L18.1899 8.23891C18.1722 8.24679 18.1545 8.25728 18.1355 8.26385C18.0967 8.27882 18.0559 8.28854 18.0144 8.29271C17.93 8.3019 17.8443 8.29271 17.764 8.29271C17.7246 8.29273 17.6852 8.29758 17.647 8.30715L17.5926 8.32421C17.5749 8.33339 17.5572 8.33995 17.5423 8.34914C17.4784 8.38353 17.4253 8.43383 17.3885 8.49479C17.36 8.55879 17.3413 8.6264 17.3327 8.69555C17.325 8.7328 17.3136 8.76927 17.2987 8.80446C17.2807 8.83994 17.2573 8.87266 17.2293 8.90156C17.17 8.96175 17.0961 9.00684 17.0143 9.03277C16.9287 9.05422 16.8395 9.05957 16.7517 9.04852C16.6747 9.03687 16.5961 9.03999 16.5204 9.05771C16.4509 9.07912 16.3875 9.11597 16.3354 9.1653C16.3104 9.18891 16.2897 9.21642 16.2741 9.24666C16.2596 9.27732 16.2487 9.30945 16.2415 9.34245C16.2303 9.41642 16.2068 9.48818 16.1721 9.55502C16.1291 9.62254 16.0674 9.67709 15.9939 9.71248C15.975 9.72212 15.9554 9.73045 15.9353 9.73741L15.8741 9.75447C15.8294 9.76325 15.7836 9.76634 15.7381 9.76365C15.6523 9.76365 15.5693 9.74266 15.4877 9.74659C15.4072 9.74956 15.3289 9.77262 15.2605 9.81351Z" fill="#20C20E" />
-            <path d="M14.587 9.52746L14.6604 9.50515C14.686 9.49908 14.7119 9.49426 14.738 9.49072C14.7917 9.48352 14.8462 9.48352 14.8999 9.49072C15.0074 9.50515 15.1081 9.54714 15.2088 9.56814C15.2583 9.57941 15.3089 9.58557 15.3598 9.58651C15.3843 9.58781 15.4088 9.58781 15.4333 9.58651C15.4576 9.58309 15.4817 9.57827 15.5054 9.57207C15.5985 9.55012 15.6839 9.50486 15.753 9.44086C15.8139 9.37341 15.8643 9.29783 15.9027 9.21648C15.9273 9.17345 15.9565 9.133 15.9898 9.09576C16.0259 9.05754 16.0682 9.02519 16.115 8.99997C16.2121 8.94726 16.3206 8.91715 16.432 8.91205C16.5452 8.91155 16.6574 8.93204 16.7626 8.97241C16.8545 9.00936 16.9529 9.02896 17.0524 9.03015C17.1476 9.02485 17.2404 8.99979 17.3246 8.95667C17.4038 8.91306 17.4688 8.84898 17.5123 8.77165C17.5572 8.68358 17.617 8.60332 17.6892 8.53415C17.767 8.46893 17.8606 8.42379 17.9613 8.40294C17.9882 8.39657 18.0154 8.39218 18.043 8.38981C18.0702 8.38981 18.0988 8.38981 18.126 8.38981C18.1818 8.39346 18.237 8.40314 18.2906 8.41868C18.3954 8.44755 18.4893 8.49347 18.5913 8.51316C18.6936 8.53488 18.8 8.53081 18.9002 8.50135L18.8267 8.52497L18.7491 8.5394C18.6959 8.54662 18.6418 8.54662 18.5886 8.5394C18.4797 8.52497 18.379 8.48429 18.2784 8.46198C18.2293 8.45079 18.1791 8.44463 18.1287 8.44361H18.0538C18.0294 8.44361 18.0049 8.45149 17.9817 8.45805C17.8888 8.4804 17.8035 8.52561 17.7341 8.58926C17.6733 8.65671 17.6229 8.73229 17.5844 8.81364C17.5601 8.85686 17.5309 8.89735 17.4974 8.93436C17.4617 8.97255 17.4199 9.00491 17.3735 9.03015C17.276 9.08329 17.167 9.11384 17.0551 9.11938C16.946 9.11855 16.8381 9.09808 16.7368 9.05902C16.6443 9.02245 16.5455 9.0033 16.4456 9.00259C16.3503 9.00681 16.2572 9.03194 16.1735 9.07608C16.1332 9.09617 16.0968 9.1228 16.066 9.15481C16.0359 9.18731 16.009 9.22245 15.9857 9.25978C15.9426 9.34645 15.8851 9.42578 15.8156 9.49466C15.7346 9.56195 15.6364 9.60724 15.5312 9.62587C15.5044 9.63224 15.4771 9.63662 15.4496 9.63899C15.4224 9.63899 15.3938 9.63899 15.3666 9.63899C15.3113 9.6359 15.2565 9.62666 15.2033 9.61144C15.0986 9.58126 15.0033 9.53533 14.9013 9.51565C14.7971 9.49435 14.6891 9.4984 14.587 9.52746Z" fill="#20C20E" />
-            <path d="M15.0333 9.02359H15.153C15.1941 9.02837 15.2344 9.03852 15.2727 9.05377C15.3448 9.0862 15.4152 9.12212 15.4836 9.16137C15.518 9.17766 15.554 9.19041 15.5911 9.19942L15.6469 9.2086C15.6659 9.2086 15.685 9.2086 15.704 9.2086C15.7782 9.2085 15.8512 9.19046 15.9163 9.15612C15.9755 9.11924 16.0294 9.07514 16.0769 9.0249C16.1052 8.99849 16.1356 8.97435 16.168 8.95273C16.2032 8.9308 16.2418 8.91441 16.2823 8.90418C16.3652 8.88222 16.4522 8.87864 16.5367 8.89368C16.6195 8.91201 16.6976 8.9459 16.7667 8.99341C16.8278 9.03529 16.8966 9.06557 16.9694 9.08264C17.0417 9.09511 17.116 9.09243 17.1871 9.07476C17.2557 9.05749 17.318 9.02204 17.3667 8.97241C17.4188 8.91756 17.4788 8.87027 17.545 8.83201C17.6198 8.79652 17.7025 8.77939 17.7858 8.78215C17.8076 8.78215 17.828 8.78215 17.8498 8.78215L17.911 8.79527C17.9513 8.80666 17.9901 8.82251 18.0266 8.84251C18.0974 8.88187 18.1627 8.93042 18.2294 8.96192C18.3012 8.99425 18.3802 9.00869 18.4593 9.00391H18.3396C18.2989 8.99973 18.2591 8.99001 18.2212 8.97504C18.1437 8.94617 18.0851 8.89893 18.009 8.86744C17.9748 8.85049 17.9387 8.83728 17.9015 8.82808C17.8824 8.82808 17.8647 8.82808 17.8457 8.8202C17.8266 8.81892 17.8076 8.81892 17.7885 8.8202C17.7149 8.81967 17.6423 8.83728 17.5776 8.87138C17.5178 8.90787 17.4633 8.95201 17.4157 9.00259C17.3886 9.02969 17.3584 9.05389 17.3259 9.07476C17.2905 9.09622 17.252 9.11258 17.2116 9.12331C17.1287 9.14522 17.0417 9.14881 16.9572 9.13381C16.874 9.11621 16.7954 9.08275 16.7259 9.0354C16.6648 8.99314 16.596 8.96241 16.5231 8.94486C16.4508 8.93238 16.3766 8.93506 16.3054 8.95273C16.2713 8.96094 16.2387 8.97422 16.2088 8.9921C16.1785 9.00962 16.1506 9.03078 16.1258 9.05508C16.0741 9.11024 16.014 9.15757 15.9476 9.19548C15.8729 9.23113 15.7901 9.24827 15.7068 9.24534C15.6859 9.24665 15.665 9.24665 15.6442 9.24534C15.623 9.24237 15.6021 9.23798 15.5816 9.23222C15.5411 9.22138 15.5022 9.2055 15.4659 9.18498C15.3952 9.14693 15.3299 9.09707 15.2646 9.06558C15.1921 9.03386 15.1127 9.01946 15.0333 9.02359Z" fill="#20C20E" />
-            <path d="M14.7135 8.67586H14.7843L14.8564 8.69161C14.9042 8.70486 14.95 8.72428 14.9924 8.74935C15.0682 8.80505 15.1409 8.86462 15.2101 8.9278C15.2464 8.95388 15.2857 8.9759 15.3271 8.99341L15.3911 9.01572C15.4142 9.01572 15.436 9.02621 15.4578 9.02884C15.5485 9.04489 15.642 9.03903 15.7299 9.01178C15.8116 8.97837 15.888 8.93421 15.9571 8.88056C15.9975 8.85483 16.0403 8.83286 16.085 8.81495C16.1326 8.79727 16.1825 8.78623 16.2333 8.78215C16.3367 8.77365 16.4407 8.78709 16.5381 8.82152C16.6304 8.85971 16.7141 8.91499 16.7844 8.98422C16.8457 9.04588 16.9193 9.09497 17.0007 9.12856C17.0835 9.15912 17.1722 9.17167 17.2606 9.1653C17.3468 9.15773 17.4296 9.1293 17.5014 9.08264C17.5776 9.02775 17.6614 8.98357 17.7504 8.95142C17.848 8.92545 17.9506 8.9223 18.0498 8.94223C18.0756 8.94223 18.0988 8.95536 18.1232 8.9606C18.1945 8.98327 18.2605 9.01891 18.3178 9.06558C18.3913 9.12594 18.4539 9.19679 18.5246 9.24666C18.6029 9.29857 18.693 9.33144 18.7872 9.34244H18.7151C18.692 9.34244 18.6675 9.33326 18.6444 9.3267C18.5964 9.31367 18.5506 9.29424 18.5083 9.26896C18.4253 9.22041 18.36 9.15218 18.2865 9.10101C18.2503 9.07451 18.211 9.05206 18.1695 9.03409C18.1487 9.02572 18.1273 9.0187 18.1056 9.01309C18.0838 9.00671 18.0615 9.00231 18.0389 8.99997C17.9504 8.98458 17.8594 8.98998 17.7736 9.01572C17.6923 9.04994 17.616 9.09404 17.5463 9.14693C17.506 9.17267 17.4631 9.19464 17.4184 9.21254C17.3709 9.23057 17.321 9.24205 17.2701 9.24666C17.1667 9.25469 17.0626 9.24081 16.9653 9.20598C16.8728 9.16873 16.7891 9.11383 16.7191 9.04458C16.658 8.98287 16.585 8.93337 16.5041 8.89893C16.4206 8.86951 16.3315 8.85744 16.2429 8.8635C16.2001 8.86571 16.1579 8.87412 16.1177 8.88843C16.0774 8.9038 16.0387 8.92269 16.002 8.94486C15.9259 8.99974 15.842 9.04392 15.753 9.07607C15.6554 9.10204 15.5528 9.10519 15.4537 9.08526C15.4273 9.07792 15.4014 9.06916 15.3761 9.05902L15.3054 9.03277C15.2614 9.01148 15.2202 8.98503 15.1829 8.95404C15.1081 8.89368 15.0469 8.82283 14.9747 8.77428C14.8975 8.72096 14.8077 8.68712 14.7135 8.67586Z" fill="#20C20E" />
-            <path d="M15.0183 8.29271L15.0809 8.31371L15.1394 8.34258C15.1774 8.36357 15.2123 8.38963 15.2428 8.41999C15.2965 8.48224 15.3469 8.54706 15.3938 8.61419C15.4197 8.64464 15.4489 8.67236 15.4809 8.69686C15.4974 8.70841 15.5147 8.71893 15.5326 8.72835C15.5504 8.73866 15.569 8.74744 15.5884 8.7546C15.6648 8.78579 15.7477 8.7997 15.8306 8.79527C15.9128 8.7838 15.9932 8.76265 16.0701 8.73229C16.1144 8.71833 16.1599 8.70822 16.2061 8.70211C16.2535 8.69622 16.3016 8.69622 16.349 8.70211C16.4426 8.71552 16.5319 8.74916 16.6102 8.80052C16.6825 8.85185 16.7424 8.91765 16.7857 8.99341C16.8233 9.06052 16.8752 9.11912 16.9381 9.1653C17.006 9.2088 17.0824 9.23828 17.1626 9.25191C17.2427 9.26351 17.3245 9.25585 17.4007 9.2296C17.4839 9.19937 17.5708 9.17995 17.6593 9.17186C17.7532 9.16908 17.8465 9.18798 17.9314 9.22697C17.9532 9.23485 17.9722 9.24797 17.9926 9.25715C18.0121 9.2691 18.0307 9.28225 18.0484 9.29652C18.0824 9.32331 18.1122 9.35468 18.1368 9.38968C18.1872 9.45792 18.2185 9.53009 18.2729 9.59044C18.3284 9.6537 18.4002 9.70168 18.4811 9.72953L18.4185 9.70854C18.3981 9.69935 18.3777 9.69148 18.3586 9.68098C18.3212 9.65918 18.2869 9.63271 18.2566 9.60225C18.1954 9.54189 18.1559 9.47104 18.1042 9.40937C18.0795 9.37832 18.0507 9.35053 18.0185 9.3267L17.9668 9.2939C17.9477 9.28602 17.93 9.27552 17.9096 9.26896C17.8338 9.23715 17.7514 9.22278 17.6688 9.22697C17.5866 9.23828 17.5061 9.25944 17.4293 9.28996C17.385 9.30384 17.3394 9.31394 17.2933 9.32014C17.2481 9.32619 17.2023 9.32619 17.1572 9.32014C17.0653 9.30907 16.9775 9.27682 16.9012 9.22612C16.825 9.17541 16.7625 9.10775 16.7191 9.02884C16.6812 8.96197 16.6293 8.90344 16.5667 8.85694C16.4995 8.81335 16.4234 8.78425 16.3435 8.77165C16.3038 8.76645 16.2635 8.76645 16.2238 8.77165C16.1834 8.77758 16.1434 8.78591 16.1041 8.79658C16.0212 8.82782 15.9342 8.84769 15.8455 8.85563C15.7516 8.85934 15.6582 8.84087 15.5734 8.80183C15.5517 8.79265 15.5326 8.78084 15.5122 8.77034L15.4564 8.73229C15.4231 8.70437 15.3934 8.67265 15.368 8.63781C15.319 8.56958 15.2863 8.49741 15.2319 8.43705C15.1747 8.37212 15.1011 8.32243 15.0183 8.29271Z" fill="#20C20E" />
-            <path d="M14.9054 7.80196L14.9734 7.84002C14.9938 7.85576 15.0156 7.86888 15.036 7.88594C15.0741 7.91916 15.1075 7.9571 15.1353 7.99879C15.1897 8.08145 15.2156 8.17462 15.2591 8.26122C15.28 8.30263 15.306 8.34141 15.3367 8.37669L15.3884 8.42524C15.4088 8.43836 15.4265 8.45411 15.4482 8.46592C15.5297 8.51928 15.6228 8.55384 15.7204 8.56696C15.8216 8.57146 15.923 8.56441 16.0224 8.54596C16.0787 8.53781 16.1356 8.53386 16.1925 8.53415C16.2496 8.53829 16.3059 8.54977 16.3599 8.56827C16.4662 8.60328 16.5626 8.66168 16.6415 8.73885C16.7118 8.81463 16.7642 8.90415 16.7953 9.00128C16.8212 9.08786 16.8678 9.16739 16.9313 9.23354C16.999 9.29977 17.0814 9.35045 17.1722 9.38181C17.2646 9.41349 17.3636 9.42294 17.4606 9.40937C17.5676 9.39209 17.6764 9.38812 17.7844 9.39756C17.8947 9.4142 17.9992 9.45605 18.0892 9.51959C18.1124 9.53402 18.1314 9.5537 18.1532 9.56945C18.1749 9.5852 18.1899 9.60619 18.2076 9.62587C18.2395 9.66601 18.2656 9.71015 18.2852 9.75709C18.3246 9.84632 18.3423 9.93686 18.3886 10.0195C18.4372 10.1048 18.5086 10.1759 18.5954 10.2255L18.5274 10.1875C18.5054 10.1736 18.4844 10.1583 18.4648 10.1416C18.4256 10.1093 18.392 10.0712 18.3654 10.0287C18.3096 9.94604 18.2852 9.85288 18.2403 9.76627C18.2199 9.72501 18.1938 9.6866 18.1627 9.65212C18.1468 9.63463 18.1295 9.61839 18.111 9.60357C18.0922 9.58811 18.0722 9.57407 18.0511 9.56158C17.97 9.50786 17.8767 9.47366 17.779 9.46185C17.6783 9.45628 17.5773 9.4629 17.4783 9.48153C17.4233 9.48838 17.3677 9.49058 17.3123 9.4881C17.2548 9.4844 17.1981 9.47337 17.1436 9.45529C17.0376 9.41898 16.9413 9.36025 16.8619 9.2834C16.7913 9.2081 16.7392 9.11839 16.7095 9.02096C16.6826 8.93478 16.6361 8.85545 16.5735 8.78871C16.5042 8.72307 16.4205 8.67329 16.3286 8.64306C16.2848 8.62696 16.2391 8.61638 16.1925 8.61157C16.1436 8.60767 16.0945 8.60767 16.0456 8.61157C15.9386 8.62883 15.8298 8.6328 15.7217 8.62338C15.6111 8.60649 15.5063 8.56468 15.4156 8.50135C15.3935 8.48617 15.3726 8.46951 15.353 8.45149C15.333 8.43398 15.3144 8.41512 15.2972 8.39506C15.2658 8.3546 15.2397 8.31052 15.2197 8.26385C15.1802 8.17462 15.1639 8.08408 15.1176 8.00141C15.0661 7.9182 14.9929 7.84944 14.9054 7.80196Z" fill="#20C20E" />
-            <path d="M15.6483 7.74685L15.6904 7.78884C15.7027 7.80459 15.7163 7.81902 15.7272 7.83477C15.7482 7.86721 15.7638 7.90263 15.7734 7.93974C15.7864 8.01185 15.7954 8.08455 15.8006 8.15756C15.8066 8.19222 15.8167 8.22611 15.8306 8.2586C15.8388 8.27434 15.8479 8.28966 15.8578 8.30452C15.8684 8.3195 15.8797 8.33395 15.8918 8.34782C15.9419 8.40556 16.0059 8.45061 16.0782 8.47904C16.1554 8.50319 16.2354 8.51817 16.3163 8.52365C16.3602 8.52967 16.4034 8.53933 16.4456 8.55252C16.4885 8.56642 16.5292 8.58584 16.5667 8.61026C16.6378 8.65783 16.6965 8.72073 16.7381 8.79396C16.7728 8.86485 16.7896 8.94263 16.7871 9.02096C16.7838 9.08994 16.7973 9.15869 16.8266 9.22173C16.8621 9.28455 16.9118 9.33882 16.9722 9.3805C17.0367 9.42278 17.1109 9.44934 17.1885 9.45792C17.2745 9.46588 17.3591 9.4844 17.4402 9.51303C17.5206 9.54678 17.5908 9.5996 17.6443 9.66655C17.6579 9.68098 17.6674 9.69935 17.6797 9.7151C17.6901 9.73237 17.6992 9.75035 17.7069 9.7689C17.7207 9.80479 17.7294 9.8423 17.7327 9.88043C17.7322 9.95274 17.7358 10.025 17.7436 10.0969C17.7576 10.1704 17.7935 10.2384 17.847 10.2925L17.8049 10.2505C17.7926 10.236 17.7804 10.2203 17.7695 10.2045C17.7482 10.1722 17.7326 10.1367 17.7232 10.0996C17.7093 10.028 17.6997 9.95571 17.6946 9.88306C17.6901 9.84767 17.68 9.81316 17.6647 9.78071C17.657 9.7647 17.6479 9.74934 17.6375 9.73478C17.6274 9.7195 17.616 9.70502 17.6035 9.69148C17.5537 9.63392 17.4902 9.58888 17.4184 9.56026C17.3411 9.53674 17.2612 9.52178 17.1803 9.51565C17.1341 9.5101 17.0886 9.50044 17.0443 9.48678C17.0017 9.47315 16.9614 9.45371 16.9245 9.42905C16.8529 9.38587 16.7946 9.32503 16.7556 9.25282C16.7167 9.1806 16.6984 9.09965 16.7027 9.01834C16.7065 8.94933 16.693 8.88046 16.6633 8.81758C16.6282 8.75446 16.5784 8.7001 16.5177 8.65881C16.4867 8.63776 16.4528 8.62098 16.417 8.60895C16.3799 8.59623 16.3416 8.587 16.3027 8.58139C16.2163 8.57344 16.1312 8.55492 16.0496 8.52628C15.9702 8.49219 15.9007 8.43994 15.8469 8.37407C15.8338 8.35808 15.8215 8.34143 15.8102 8.32421C15.8006 8.30715 15.7911 8.28878 15.783 8.27041C15.7691 8.23455 15.7604 8.19702 15.7571 8.15887C15.7571 8.08408 15.7571 8.01322 15.7476 7.94237C15.7342 7.86949 15.6998 7.80172 15.6483 7.74685Z" fill="#20C20E" />
-            <path d="M15.8605 7.44374L15.8959 7.49623C15.9061 7.51452 15.9152 7.53336 15.9231 7.55265C15.939 7.59113 15.9477 7.63198 15.949 7.67337C15.9451 7.75333 15.9369 7.83305 15.9245 7.91219C15.9218 7.95042 15.9245 7.98883 15.9326 8.02634C15.9326 8.04471 15.9435 8.06308 15.9503 8.08145C15.9916 8.17417 16.0608 8.2528 16.149 8.30715C16.2279 8.34748 16.3115 8.37875 16.398 8.40031C16.445 8.41397 16.4906 8.43199 16.534 8.45411C16.5772 8.47683 16.6166 8.50557 16.651 8.5394C16.716 8.60349 16.7628 8.68254 16.7871 8.76903C16.8083 8.85176 16.8083 8.93823 16.7871 9.02096C16.7688 9.09383 16.7688 9.16986 16.7871 9.24272C16.8095 9.31689 16.8506 9.38453 16.9068 9.43955C16.9672 9.49767 17.0415 9.54049 17.1232 9.5642C17.2152 9.58839 17.3035 9.62412 17.3858 9.67049C17.4645 9.7208 17.5282 9.78998 17.5708 9.87125C17.5806 9.89089 17.5892 9.91104 17.5967 9.93161C17.6035 9.9526 17.6089 9.9736 17.6144 9.99328C17.621 10.0368 17.621 10.081 17.6144 10.1245C17.5958 10.2015 17.5817 10.2795 17.5722 10.3581C17.5693 10.4392 17.5921 10.5193 17.6375 10.5877L17.6008 10.5339C17.5926 10.5155 17.5817 10.4971 17.5735 10.4775C17.5589 10.4387 17.5506 10.398 17.5491 10.3567C17.5517 10.2772 17.5594 10.1979 17.5722 10.1192C17.5754 10.0811 17.5731 10.0427 17.5654 10.0051C17.5607 9.98623 17.5543 9.96779 17.5463 9.94998C17.5386 9.93143 17.5295 9.91344 17.5191 9.89618C17.4784 9.82224 17.4184 9.75983 17.345 9.7151C17.2661 9.67487 17.1825 9.64402 17.096 9.62325C17.0491 9.60855 17.0036 9.59011 16.9599 9.56814C16.9163 9.54543 16.8765 9.51669 16.8415 9.48285C16.7776 9.41853 16.7309 9.34016 16.7055 9.25453C16.684 9.17138 16.684 9.08443 16.7055 9.00128C16.7231 8.92834 16.7231 8.85247 16.7055 8.77953C16.682 8.70551 16.6405 8.638 16.5844 8.5827C16.5559 8.55399 16.5234 8.52925 16.4878 8.50922C16.4495 8.48878 16.4094 8.47166 16.368 8.45805C16.2761 8.43361 16.1879 8.39789 16.1054 8.35176C16.0276 8.30084 15.9644 8.23178 15.9217 8.151C15.9115 8.13137 15.9025 8.11122 15.8945 8.09064C15.8877 8.07046 15.8822 8.04986 15.8782 8.02897C15.8706 7.98554 15.8706 7.94118 15.8782 7.89775C15.897 7.82069 15.9115 7.74273 15.9217 7.66419C15.9236 7.58657 15.9024 7.51005 15.8605 7.44374Z" fill="#20C20E" />
-            <path d="M16.3871 7.46605L16.4048 7.51985C16.4048 7.53822 16.4048 7.55659 16.4143 7.57496C16.4167 7.61202 16.4131 7.64923 16.4034 7.68518C16.3779 7.75348 16.3489 7.82049 16.3163 7.88594C16.3029 7.91823 16.2946 7.9523 16.2918 7.98698C16.2905 8.00401 16.2905 8.02112 16.2918 8.03815C16.2929 8.05594 16.2961 8.07356 16.3014 8.09064C16.3177 8.16052 16.354 8.2246 16.4061 8.27565C16.4651 8.32587 16.5305 8.36864 16.6007 8.40294C16.6383 8.42474 16.6739 8.44977 16.7068 8.47773C16.7382 8.50715 16.7656 8.5402 16.7885 8.57614C16.8301 8.64556 16.8521 8.72427 16.8524 8.80446C16.8469 8.8822 16.8216 8.95744 16.7789 9.02359C16.7411 9.08137 16.7183 9.14708 16.7123 9.21517C16.7124 9.28363 16.7307 9.35091 16.7653 9.41068C16.8036 9.47402 16.8585 9.52652 16.9245 9.56289C17.0006 9.60455 17.0706 9.65573 17.1327 9.7151C17.1885 9.77549 17.2262 9.84945 17.2416 9.92898C17.2458 9.94751 17.2481 9.96642 17.2484 9.98541C17.2497 10.0046 17.2497 10.0239 17.2484 10.0431C17.242 10.0807 17.2306 10.1173 17.2143 10.1521C17.1817 10.2203 17.1354 10.2833 17.1109 10.3463C17.0885 10.4155 17.0885 10.4896 17.1109 10.5588L17.0932 10.505C17.0932 10.488 17.0932 10.4696 17.0851 10.4512C17.0812 10.4142 17.0849 10.3767 17.096 10.341C17.1204 10.2721 17.149 10.2046 17.1817 10.1389C17.1951 10.1066 17.2034 10.0726 17.2062 10.0379C17.2075 10.0209 17.2075 10.0038 17.2062 9.98672C17.2062 9.96966 17.2062 9.9526 17.1967 9.93554C17.1792 9.86569 17.1431 9.80145 17.0919 9.74922C17.0319 9.69795 16.9656 9.65389 16.8946 9.618C16.8573 9.59629 16.8222 9.57125 16.7898 9.54321C16.7571 9.51506 16.7295 9.48179 16.7082 9.44479C16.6658 9.3757 16.6437 9.29677 16.6442 9.21648C16.6493 9.13884 16.6741 9.06359 16.7164 8.99735C16.7548 8.93984 16.7777 8.87399 16.783 8.80577C16.7833 8.73726 16.765 8.66986 16.73 8.61026C16.713 8.57873 16.6904 8.55031 16.6633 8.52628C16.6348 8.50076 16.6038 8.47792 16.5708 8.45805C16.4966 8.41746 16.4281 8.36809 16.3667 8.31108C16.3106 8.25137 16.2729 8.17776 16.2578 8.09851C16.2539 8.07945 16.2511 8.06017 16.2497 8.04078C16.2497 8.0211 16.2497 8.00141 16.2497 7.98304C16.2549 7.94539 16.2659 7.90869 16.2823 7.87413C16.3205 7.81132 16.355 7.74648 16.3857 7.67993C16.4096 7.61058 16.4101 7.53568 16.3871 7.46605Z" fill="#20C20E" />
-            <path d="M16.8551 7.58939V7.64057C16.8551 7.65631 16.8551 7.67337 16.8551 7.69043C16.8467 7.72337 16.8329 7.75481 16.8143 7.78359C16.7719 7.83837 16.7265 7.89093 16.6783 7.94105C16.6583 7.96648 16.6419 7.99426 16.6293 8.02372C16.6235 8.03821 16.6189 8.05313 16.6157 8.06833C16.6137 8.08402 16.6137 8.09988 16.6157 8.11557C16.6112 8.17928 16.6254 8.24291 16.6565 8.29927C16.6945 8.35685 16.7403 8.40936 16.7925 8.45542C16.8192 8.48295 16.8438 8.51234 16.866 8.54334C16.8864 8.57516 16.902 8.60962 16.9123 8.64569C16.932 8.71526 16.932 8.78868 16.9123 8.85826C16.8853 8.9258 16.8419 8.98611 16.7857 9.03409C16.7364 9.07624 16.6978 9.12876 16.6728 9.18761C16.6527 9.2482 16.6509 9.31304 16.6675 9.3746C16.6841 9.43615 16.7185 9.49189 16.7667 9.53533C16.8199 9.58598 16.8656 9.64337 16.9028 9.70591C16.937 9.77018 16.9507 9.84277 16.9422 9.91455C16.9422 9.93161 16.9422 9.94867 16.9327 9.96572C16.9232 9.98278 16.9204 9.99984 16.9136 10.0156C16.8984 10.0476 16.8782 10.0773 16.8538 10.1035C16.8062 10.156 16.7476 10.1967 16.7082 10.2505C16.6676 10.3043 16.6453 10.3689 16.6442 10.4355V10.3856C16.6428 10.369 16.6428 10.3523 16.6442 10.3358C16.6518 10.3021 16.6656 10.2701 16.6851 10.2413C16.7275 10.1865 16.7729 10.1339 16.8211 10.0838C16.8417 10.0586 16.8587 10.0308 16.8715 10.0012C16.8715 9.98672 16.881 9.97097 16.8851 9.95654C16.8892 9.9421 16.8851 9.92505 16.8851 9.9093C16.8897 9.84543 16.875 9.78163 16.8429 9.7256C16.8049 9.66796 16.7592 9.61545 16.7068 9.56945C16.6787 9.54315 16.654 9.51363 16.6334 9.48153C16.6121 9.45013 16.5965 9.41554 16.5871 9.37919C16.5669 9.30967 16.5669 9.23613 16.5871 9.16662C16.6145 9.09933 16.6579 9.03913 16.7136 8.99079C16.763 8.94914 16.8016 8.89704 16.8266 8.83857C16.8456 8.77949 16.8484 8.71663 16.8347 8.65618C16.8271 8.62569 16.8147 8.59649 16.798 8.56958C16.7792 8.54112 16.7578 8.51433 16.734 8.48954C16.6803 8.43938 16.6345 8.38191 16.598 8.31896C16.5648 8.25436 16.5516 8.18195 16.5599 8.11032C16.5617 8.09306 16.5649 8.07595 16.5694 8.05915C16.5741 8.04215 16.58 8.02548 16.5871 8.00929C16.6026 7.97775 16.6227 7.94857 16.647 7.92268C16.6987 7.8759 16.7478 7.8264 16.7939 7.77441C16.8336 7.7203 16.855 7.65568 16.8551 7.58939Z" fill="#20C20E" />
-            <path d="M17.4171 7.22199L17.3994 7.28628C17.3899 7.30728 17.3817 7.32827 17.3708 7.34795C17.3482 7.38865 17.3188 7.42546 17.2837 7.45686C17.213 7.52247 17.1218 7.56577 17.0497 7.62089C17.0136 7.64742 16.9811 7.67829 16.9531 7.71274C16.9405 7.73034 16.9291 7.74875 16.9191 7.76785C16.9091 7.78649 16.9004 7.80578 16.8932 7.82558C16.8648 7.90499 16.8605 7.9906 16.881 8.07227C16.912 8.15719 16.955 8.23756 17.0089 8.31108C17.0342 8.35289 17.0556 8.39681 17.0728 8.4423C17.0881 8.48896 17.0955 8.53771 17.0946 8.58664C17.0918 8.68273 17.0617 8.77623 17.0075 8.85694C16.9465 8.93575 16.8679 9.00033 16.7776 9.0459C16.6974 9.08535 16.6273 9.14135 16.5721 9.20992C16.5253 9.2792 16.4989 9.3595 16.4959 9.44217C16.4936 9.52711 16.5178 9.61077 16.5653 9.6823C16.6204 9.76306 16.6622 9.85159 16.6891 9.94473C16.709 10.0376 16.6995 10.134 16.6619 10.2216C16.6534 10.2434 16.643 10.2645 16.6306 10.2846C16.6178 10.3051 16.6037 10.3248 16.5885 10.3436C16.5559 10.3799 16.5184 10.4117 16.4769 10.4381C16.3952 10.4919 16.3027 10.5286 16.2306 10.5877C16.1581 10.6435 16.1053 10.7196 16.0796 10.8055L16.0973 10.7412C16.1049 10.7201 16.114 10.6995 16.1245 10.6795C16.1476 10.6388 16.1774 10.602 16.2129 10.5706C16.2876 10.5116 16.3657 10.4569 16.4469 10.4066C16.4823 10.3797 16.5143 10.3489 16.5422 10.3148C16.5544 10.2964 16.5667 10.2793 16.5776 10.261C16.5871 10.2418 16.5953 10.2221 16.6021 10.2019C16.6316 10.1227 16.6363 10.037 16.6157 9.95523C16.5853 9.87056 16.5431 9.79023 16.4905 9.71641C16.4644 9.67501 16.4429 9.63102 16.4265 9.5852C16.4106 9.53868 16.4032 9.48983 16.4048 9.44086C16.4066 9.3448 16.4363 9.25117 16.4905 9.17055C16.5517 9.09193 16.6302 9.02739 16.7204 8.9816C16.8006 8.94215 16.8707 8.88615 16.9259 8.81758C16.9724 8.74854 16.9991 8.66887 17.0034 8.58664C17.0051 8.54546 17.0001 8.50428 16.9885 8.46461C16.9738 8.42385 16.9556 8.38435 16.9341 8.34651C16.8792 8.26562 16.8374 8.17712 16.8102 8.08408C16.7899 7.99094 16.7989 7.89407 16.8361 7.8059C16.8443 7.78359 16.8579 7.76391 16.8674 7.74292L16.9096 7.68387C16.9425 7.64796 16.9799 7.61621 17.0211 7.58939C17.1041 7.53559 17.1953 7.49885 17.2674 7.44112C17.3389 7.38419 17.3911 7.30785 17.4171 7.22199Z" fill="#20C20E" />
-            <path d="M19.0981 9.51627C19.3538 8.40481 18.5097 7.27882 17.2126 7.00129C15.9156 6.72375 14.6569 7.39978 14.4011 8.51124C14.1454 9.6227 14.9896 10.7487 16.2866 11.0262C17.5836 11.3038 18.8424 10.6277 19.0981 9.51627Z" fill="url(#paint0_radial_156_10958)" />
-            <path d="M16.6783 11.1034C17.8686 11.1034 18.8335 10.1781 18.8335 9.03671C18.8335 7.89533 17.8686 6.97005 16.6783 6.97005C15.4879 6.97005 14.523 7.89533 14.523 9.03671C14.523 10.1781 15.4879 11.1034 16.6783 11.1034Z" fill="url(#paint1_radial_156_10958)" />
-            <path d="M16.6538 10.3239C17.4067 10.3239 18.0171 9.73882 18.0171 9.01703C18.0171 8.29524 17.4067 7.71011 16.6538 7.71011C15.9008 7.71011 15.2904 8.29524 15.2904 9.01703C15.2904 9.73882 15.9008 10.3239 16.6538 10.3239Z" fill="url(#paint2_radial_156_10958)" />
-            <path d="M17.5123 9.18279C17.6009 8.79792 17.3087 8.40804 16.8597 8.31197C16.4108 8.2159 15.975 8.45002 15.8865 8.83489C15.7979 9.21976 16.0901 9.60964 16.5391 9.70571C16.988 9.80178 17.4238 9.56766 17.5123 9.18279Z" fill="url(#paint3_radial_156_10958)" />
-            <path d="M17.398 10.1849C17.7039 10.1849 17.9518 9.9575 17.9518 9.67705C17.9518 9.39659 17.7039 9.16924 17.398 9.16924C17.0922 9.16924 16.8443 9.39659 16.8443 9.67705C16.8443 9.9575 17.0922 10.1849 17.398 10.1849Z" fill="url(#paint4_radial_156_10958)" />
-            <path d="M1.80247 9.07214C0.667705 15.3128 0.65818 18.5617 0.00099538 23.7631C-0.00379774 23.8121 0.00881546 23.8611 0.0367614 23.9022C0.0647074 23.9433 0.106323 23.974 0.154767 23.9893C0.20321 24.0046 0.255599 24.0035 0.30332 23.9862C0.351041 23.969 0.391254 23.9366 0.417349 23.8943C1.21332 21.333 2.73042 19.1496 4.2693 17.0055C6.13472 14.2604 9.49004 13.1713 12.7406 13.1294C15.0169 12.9181 16.1599 13.4902 18.1899 13.0742C20.8418 12.4313 22.797 9.23353 24.1427 6.79947C25.1836 4.35228 27 2.60054 27 0.0116428C26.9048 -0.13532 25.8761 1.15191 24.823 1.45502C18.0647 3.06899 9.09954 -0.221922 4.39583 2.76063C3.18351 3.75262 2.52769 4.8719 2.62701 5.5726C2.69777 5.6159 2.82022 5.53717 2.8556 5.51617C3.56313 5.17763 4.50741 4.8286 4.50741 4.8286C7.83551 3.23169 13.0631 2.58086 17.1082 2.57692C17.9872 2.56774 24.8638 2.67009 26.4571 0.936719C22.3752 9.99065 11.2289 0.974771 1.0446 21.2031C1.14393 20.6599 2.30454 14.3969 3.76586 11.5377C5.02445 9.13775 4.80947 9.04458 6.40141 6.86245C6.94566 5.78779 2.53177 8.0001 2.2855 8.32421C1.89635 8.74803 2.00521 8.66537 1.80247 9.07214ZM19.9628 10.198C18.8349 11.6728 17.1558 11.787 14.7067 11.6256C12.8889 11.6597 10.1227 11.808 8.192 12.9378C12.2099 8.02766 16.7463 10.3909 21.635 6.92281C21.635 6.92281 20.7044 9.35819 19.9628 10.198Z" fill="black" />
-            <defs>
-              <radialGradient id="paint0_radial_156_10958" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(13.1263 11.3662) rotate(94.8246) scale(11.9786 13.14)">
-                <stop offset="0.07" stop-color="#00C8DD" stopOpacity="0.97" />
-                <stop offset="0.19" stop-color="#00D0E2" stopOpacity="0.97" />
-                <stop offset="0.37" stop-color="#00E4EE" stopOpacity="0.99" />
-                <stop offset="0.56" stop-color="#00FFFF" />
-                <stop offset="0.6" stop-color="#00F6F6" />
-                <stop offset="0.67" stop-color="#00DDDE" />
-                <stop offset="0.76" stop-color="#00B5B6" />
-                <stop offset="0.86" stop-color="#007D7F" />
-                <stop offset="0.97" stop-color="#003739" />
-                <stop offset="1" stop-color="#002628" />
-              </radialGradient>
-              <radialGradient id="paint1_radial_156_10958" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(12.7336 12.064) scale(15.8564 14.1751)">
-                <stop offset="0.01" stop-color="#4DFFFF" stopOpacity="0.66" />
-                <stop offset="0.09" stop-color="#41FFFF" stopOpacity="0.72" />
-                <stop offset="0.23" stop-color="#21FFFF" stopOpacity="0.86" />
-                <stop offset="0.36" stop-color="#00FFFF" />
-                <stop offset="0.41" stop-color="#05FFFF" stopOpacity="0.98" />
-                <stop offset="0.47" stop-color="#14FFFF" stopOpacity="0.92" />
-                <stop offset="0.53" stop-color="#2CFFFF" stopOpacity="0.83" />
-                <stop offset="0.59" stop-color="#4EFFFF" stopOpacity="0.7" />
-                <stop offset="0.66" stop-color="#79FFFF" stopOpacity="0.53" />
-                <stop offset="0.73" stop-color="#AEFFFF" stopOpacity="0.32" />
-                <stop offset="0.8" stop-color="#ECFFFF" stopOpacity="0.08" />
-                <stop offset="0.82" stop-color="white" stopOpacity="0" />
-              </radialGradient>
-              <radialGradient id="paint2_radial_156_10958" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(12.275 12.108) scale(13.424 12.0043)">
-                <stop stop-color="#00FFFF" />
-                <stop offset="0.28" stop-color="#00FFFF" stopOpacity="0.99" />
-                <stop offset="0.4" stop-color="#00FFFF" stopOpacity="0.96" />
-                <stop offset="0.5" stop-color="#00FFFF" stopOpacity="0.91" />
-                <stop offset="0.58" stop-color="#00FFFF" stopOpacity="0.84" />
-                <stop offset="0.65" stop-color="#00FFFF" stopOpacity="0.75" />
-                <stop offset="0.71" stop-color="#00FFFF" stopOpacity="0.64" />
-                <stop offset="0.77" stop-color="#00FFFF" stopOpacity="0.5" />
-                <stop offset="0.82" stop-color="#00FFFF" stopOpacity="0.34" />
-                <stop offset="0.87" stop-color="#00FFFF" stopOpacity="0.17" />
-                <stop offset="0.91" stop-color="#00FFFF" stopOpacity="0" />
-              </radialGradient>
-              <radialGradient id="paint3_radial_156_10958" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(12.4302 10.1745) rotate(94.8229) scale(11.9807 13.1376)">
-                <stop offset="0.02" stopOpacity="0.97" />
-                <stop offset="0.33" stopOpacity="0.98" />
-                <stop offset="1" />
-              </radialGradient>
-              <radialGradient id="paint4_radial_156_10958" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(10.6889 9.92248) scale(13.5481 12.062)">
-                <stop offset="0.33" stop-color="#F6F6F6" />
-                <stop offset="0.38" stop-color="#EFF6F6" />
-                <stop offset="0.46" stop-color="#DAF7F7" />
-                <stop offset="0.55" stop-color="#B9F8F8" />
-                <stop offset="0.66" stop-color="#8BFAFA" />
-                <stop offset="0.77" stop-color="#4FFCFC" />
-                <stop offset="0.9" stop-color="#08FFFF" />
-                <stop offset="0.91" stop-color="#00FFFF" />
-                <stop offset="1" stop-color="#00FFFF" stopOpacity="0.3" />
-              </radialGradient>
-            </defs>
-          </svg>
-
+          <img src={fridaySearch} alt="star" />
           <input
-
-           
-              type="text"
+          type="text"
               className="search-bar-NC"
               onKeyUp={(e) => {
                 searchRef.current = e.target.value;
@@ -326,28 +143,13 @@ files.</p>
                 </button>  
           <div>
           <button className='btn-icon'>
-          <svg width="23" height="23" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M17.627 12.5015C17.627 15.608 15.1085 18.1265 12.002 18.1265C8.89539 18.1265 6.37695 15.608 6.37695 12.5015C6.37695 9.3949 8.89539 6.87646 12.002 6.87646C15.1085 6.87646 17.627 9.3949 17.627 12.5015Z" fill="black" fill-opacity="0.1"/>
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M12.002 6.12646C12.002 6.12646 13.2984 6.12646 14.4836 6.62775C14.4836 6.62775 15.6278 7.1117 16.5098 7.99366C16.5098 7.99366 17.3917 8.87562 17.8757 10.0198C17.8757 10.0198 18.377 11.205 18.377 12.5015C18.377 12.5015 18.377 13.798 17.8757 14.9831C17.8757 14.9831 17.3917 16.1273 16.5098 17.0093C16.5098 17.0093 15.6278 17.8912 14.4836 18.3752C14.4836 18.3752 13.2984 18.8765 12.002 18.8765C12.002 18.8765 10.7055 18.8765 9.52029 18.3752C9.52029 18.3752 8.37611 17.8912 7.49415 17.0093C7.49415 17.0093 6.61219 16.1273 6.12824 14.9831C6.12824 14.9831 5.62695 13.798 5.62695 12.5015C5.62695 12.5015 5.62695 11.205 6.12824 10.0198C6.12824 10.0198 6.61219 8.87562 7.49415 7.99366C7.49415 7.99366 8.37611 7.1117 9.52029 6.62775C9.52029 6.62775 10.7055 6.12646 12.002 6.12646ZM12.002 7.62646C12.002 7.62646 9.98266 7.62646 8.55481 9.05432C8.55481 9.05432 7.12695 10.4822 7.12695 12.5015C7.12695 12.5015 7.12695 14.5208 8.55481 15.9486C8.55481 15.9486 9.98266 17.3765 12.002 17.3765C12.002 17.3765 14.0212 17.3765 15.4491 15.9486C15.4491 15.9486 16.877 14.5208 16.877 12.5015C16.877 12.5015 16.877 10.4822 15.4491 9.05432C15.4491 9.05432 14.0212 7.62646 12.002 7.62646Z" fill="#1C1C1C"/>
-            <path d="M12.752 3.87646V2.00146C12.752 1.58725 12.4162 1.25146 12.002 1.25146C11.5877 1.25146 11.252 1.58725 11.252 2.00146V3.87646C11.252 4.29068 11.5877 4.62646 12.002 4.62646C12.4162 4.62646 12.752 4.29068 12.752 3.87646Z" fill="#1C1C1C"/>
-            <path d="M5.10723 4.54633C4.96658 4.40568 4.77582 4.32666 4.5769 4.32666C4.37799 4.32666 4.18723 4.40568 4.04657 4.54633C3.90592 4.68698 3.8269 4.87775 3.8269 5.07666C3.8269 5.27557 3.90592 5.46634 4.04657 5.60699L5.36845 6.92887C5.5091 7.06952 5.69987 7.14854 5.89878 7.14854C6.09769 7.14854 6.28846 7.06952 6.42911 6.92887C6.56976 6.78821 6.64878 6.59745 6.64878 6.39854C6.64878 6.19962 6.56976 6.00886 6.42911 5.86821L5.10723 4.54633Z" fill="#1C1C1C"/>
-            <path d="M3.37695 11.7515H1.50195C1.08774 11.7515 0.751953 12.0873 0.751953 12.5015C0.751953 12.9157 1.08774 13.2515 1.50195 13.2515H3.37695C3.79117 13.2515 4.12695 12.9157 4.12695 12.5015C4.12695 12.0873 3.79117 11.7515 3.37695 11.7515Z" fill="#1C1C1C"/>
-            <path d="M4.04657 19.396C3.90592 19.5367 3.8269 19.7275 3.8269 19.9264C3.8269 19.9393 3.82724 19.9522 3.82791 19.9652C3.83751 20.1504 3.91541 20.3255 4.04657 20.4567C4.18723 20.5973 4.37799 20.6764 4.5769 20.6764C4.77582 20.6764 4.96658 20.5973 5.10723 20.4567L6.42911 19.1348C6.56976 18.9942 6.64878 18.8034 6.64878 18.6045C6.64878 18.4056 6.56976 18.2148 6.42911 18.0742C6.28846 17.9335 6.09769 17.8545 5.89878 17.8545C5.88584 17.8545 5.8729 17.8548 5.85997 17.8555C5.67473 17.8651 5.49961 17.943 5.36845 18.0742L4.04657 19.396Z" fill="#1C1C1C"/>
-            <path d="M11.252 21.1265V23.0015C11.252 23.4157 11.5877 23.7515 12.002 23.7515C12.4162 23.7515 12.752 23.4157 12.752 23.0015V21.1265C12.752 20.7123 12.4162 20.3765 12.002 20.3765C11.5877 20.3765 11.252 20.7123 11.252 21.1265Z" fill="#1C1C1C"/>
-            <path d="M17.5747 19.1349L18.8965 20.4567C19.0372 20.5973 19.2279 20.6764 19.4269 20.6764C19.6258 20.6764 19.8165 20.5973 19.9572 20.4567C20.0978 20.316 20.1769 20.1253 20.1769 19.9264C20.1769 19.7275 20.0978 19.5367 19.9572 19.396L18.6354 18.0743C18.4947 17.9335 18.3039 17.8545 18.105 17.8545C17.9061 17.8545 17.7153 17.9335 17.5747 18.0742C17.434 18.2148 17.355 18.4056 17.355 18.6045C17.355 18.8034 17.434 18.9942 17.5747 19.1349Z" fill="#1C1C1C"/>
-            <path d="M20.627 13.2515H22.502C22.9162 13.2515 23.252 12.9157 23.252 12.5015C23.252 12.0873 22.9162 11.7515 22.502 11.7515H20.627C20.2127 11.7515 19.877 12.0873 19.877 12.5015C19.877 12.9157 20.2127 13.2515 20.627 13.2515Z" fill="#1C1C1C"/>
-            <path d="M19.9572 5.60699C20.0978 5.46634 20.1769 5.27557 20.1769 5.07666C20.1769 4.87775 20.0978 4.68698 19.9572 4.54633C19.8165 4.40568 19.6258 4.32666 19.4269 4.32666C19.2279 4.32666 19.0372 4.40568 18.8965 4.54633L17.5747 5.86812C17.434 6.00886 17.355 6.19962 17.355 6.39854C17.355 6.59745 17.434 6.78821 17.5747 6.92887C17.7153 7.06952 17.9061 7.14854 18.105 7.14854C18.3039 7.14854 18.4947 7.06952 18.6353 6.92887L19.9572 5.60699Z" fill="#1C1C1C"/>
-</svg>
+         <img src={sun} alt="sun" />
 </button>
 
         </div>
         <div>
           <button className='btn-icon' onClick={notification}>
-            <svg width="23" height="23" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M5.2688 9.75009C5.26752 8.86098 5.4422 7.98033 5.78278 7.15901C6.12335 6.33765 6.62305 5.59181 7.25293 4.9644C7.883 4.33699 8.63098 3.84041 9.45367 3.50327C10.2766 3.16617 11.1578 2.99515 12.047 3.00009C15.7595 3.02824 18.7313 6.11261 18.7313 9.83446V10.5001C18.7313 13.8564 19.4344 15.8064 20.0531 16.8751C20.1188 16.9889 20.1536 17.118 20.1536 17.2495C20.1538 17.3809 20.1194 17.51 20.0538 17.6239C19.9883 17.7379 19.894 17.8326 19.7805 17.8987C19.6667 17.9647 19.5377 17.9996 19.4064 18.0001H4.59387C4.4624 17.9996 4.33331 17.9647 4.21979 17.8987C4.10608 17.8326 4.01178 17.7379 3.94623 17.6239C3.88068 17.51 3.84625 17.3809 3.84644 17.2495C3.84662 17.118 3.88123 16.9889 3.94696 16.8751C4.56567 15.8064 5.2688 13.8564 5.2688 10.5001V9.75009Z" fill="black" fill-opacity="0.1"/>
-              <path fill-rule="evenodd" clip-rule="evenodd" d="M4.51879 10.5001V9.75008C4.51879 9.75008 4.51669 8.25437 5.08994 6.87174C5.08994 6.87174 5.66319 5.4891 6.72376 4.43293C6.72376 4.43293 7.78433 3.37676 9.16934 2.80926C9.16934 2.80926 10.5544 2.24176 12.0511 2.25009C12.0511 2.25009 13.573 2.26162 14.9589 2.86961C14.9589 2.86961 16.2922 3.45457 17.3187 4.50894C17.3187 4.50894 18.3411 5.55919 18.9019 6.91345C18.9019 6.91345 19.4813 8.31283 19.4813 9.83445V10.5001C19.4813 10.5001 19.4813 14.3904 20.7022 16.4993C20.7022 16.4993 20.9033 16.8475 20.9037 17.2488C20.9037 17.2488 20.904 17.65 20.704 17.9978C20.704 17.9978 20.504 18.3456 20.1571 18.5471C20.1571 18.5471 19.8102 18.7487 19.4089 18.7501L4.59379 18.7501C4.59379 18.7501 4.18993 18.7487 3.843 18.5471C3.843 18.5471 3.49607 18.3456 3.29608 17.9978C3.29608 17.9978 3.09608 17.65 3.09644 17.2488C3.09644 17.2488 3.09679 16.8475 3.2974 16.5001C3.2974 16.5001 4.51879 14.3904 4.51879 10.5001ZM6.01879 10.5001C6.01879 10.5001 6.01879 14.7933 4.59644 17.2501L19.4036 17.2501C19.4036 17.2501 17.9813 14.7926 17.9813 10.5001V9.83445C17.9813 9.83445 17.9813 7.33991 16.2439 5.55528C16.2439 5.55528 14.5046 3.76872 12.0427 3.75007C12.0427 3.75007 10.8457 3.74341 9.73807 4.19726C9.73807 4.19726 8.63041 4.65112 7.78222 5.49579C7.78222 5.49579 6.93403 6.34046 6.47557 7.44623C6.47557 7.44623 6.01711 8.55199 6.01879 9.75008V10.5001Z" fill="#1C1C1C"/>
-              <path d="M15.75 18.75V18C15.75 17.5858 15.4142 17.25 15 17.25C14.5858 17.25 14.25 17.5858 14.25 18V18.75C14.25 19.682 13.591 20.341 13.591 20.341C12.932 21 12 21 12 21C11.068 21 10.409 20.341 10.409 20.341C9.75 19.682 9.75 18.75 9.75 18.75V18C9.75 17.5858 9.41421 17.25 9 17.25C8.58579 17.25 8.25 17.5858 8.25 18L8.25 18.75C8.25 20.3033 9.34835 21.4016 9.34835 21.4016C10.4467 22.5 12 22.5 12 22.5C13.5533 22.5 14.6516 21.4017 14.6516 21.4017C15.75 20.3033 15.75 18.75 15.75 18.75Z" fill="#1C1C1C"/>
-            </svg>
+            <img src={bell} alt="bell" />
           </button>
           {/* NOTIFICATION BUTTON DATA---------------------------------------------------------------- */}
           {isshow && (
@@ -360,11 +162,7 @@ files.</p>
         <div  onClick={() => navigate(ROUTES.member)}  onMouseOver={handleMouseOver}
           onMouseOut={handleMouseOut}>
           <button className='btn-icon member-icon hover-member'>
-            <svg width="23" height="23" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path d="M18 9C18 12.3137 15.3137 15 12 15C8.68634 15 6 12.3137 6 9C6 5.68634 8.68634 3 12 3C15.3137 3 18 5.68634 18 9Z" fill="black" fill-opacity="0.1"/>
-              <path fill-rule="evenodd" clip-rule="evenodd" d="M12 2.25C12 2.25 13.3728 2.25 14.6276 2.78076C14.6276 2.78076 15.8391 3.29317 16.773 4.22703C16.773 4.22703 17.7068 5.1609 18.2192 6.37236C18.2192 6.37236 18.75 7.62721 18.75 9C18.75 9 18.75 10.3728 18.2192 11.6276C18.2192 11.6276 17.7068 12.8391 16.773 13.773C16.773 13.773 15.8391 14.7068 14.6276 15.2192C14.6276 15.2192 13.3728 15.75 12 15.75C12 15.75 10.6272 15.75 9.37237 15.2192C9.37237 15.2192 8.16089 14.7068 7.22703 13.773C7.22703 13.773 6.29317 12.8391 5.78076 11.6276C5.78076 11.6276 5.25 10.3728 5.25 9C5.25 9 5.25 7.62721 5.78076 6.37236C5.78076 6.37236 6.29317 5.16089 7.22703 4.22703C7.22703 4.22703 8.16089 3.29317 9.37237 2.78076C9.37237 2.78076 10.6272 2.25 12 2.25ZM12 3.75C12 3.75 9.82538 3.75 8.28769 5.28769C8.28769 5.28769 6.75 6.82538 6.75 9C6.75 9 6.75 11.1746 8.28769 12.7123C8.28769 12.7123 9.82538 14.25 12 14.25C12 14.25 14.1746 14.25 15.7123 12.7123C15.7123 12.7123 17.25 11.1746 17.25 9C17.25 9 17.25 6.82538 15.7123 5.28769C15.7123 5.28769 14.1746 3.75 12 3.75Z" fill="#1C1C1C"/>
-              <path d="M16.8754 17.0555C19.138 18.362 20.4442 20.6248 20.4442 20.6248C20.5436 20.7971 20.7074 20.9228 20.8996 20.9743C20.9629 20.9913 21.0282 20.9999 21.0938 20.9999C21.1016 20.9999 21.1095 20.9997 21.1173 20.9995C21.2409 20.9956 21.3616 20.9612 21.4687 20.8994C21.7008 20.7654 21.8438 20.5178 21.8438 20.2499C21.8438 20.2383 21.8435 20.2267 21.8429 20.2151C21.8374 20.0954 21.8032 19.9787 21.7433 19.8749C20.2362 17.264 17.6254 15.7565 17.6254 15.7565C15.0147 14.249 12 14.249 12 14.249C8.98529 14.249 6.37455 15.7565 6.37455 15.7565C3.76401 17.2639 2.25693 19.8745 2.25693 19.8745L2.2567 19.8749C2.19089 19.9889 2.15625 20.1182 2.15625 20.2499L2.15637 20.2634C2.15748 20.3244 2.16602 20.3851 2.18182 20.444C2.23332 20.6362 2.35904 20.8 2.53131 20.8994C2.64531 20.9652 2.77462 20.9999 2.90625 20.9999C2.92013 20.9999 2.93401 20.9995 2.94787 20.9987C3.20063 20.9847 3.42925 20.844 3.5558 20.6248C4.86197 18.362 7.12461 17.0555 7.12461 17.0555C9.38726 15.749 12 15.749 12 15.749C14.6127 15.749 16.8754 17.0555 16.8754 17.0555Z" fill="#1C1C1C"/>
-            </svg>
+           <img src={user} alt="user" />
           </button>
